Fixed balph files and renamed folder
Balph files had a bunch of HTML and yuck in them. Fixed it with the code
from the submission thread and renamed the folder to balphagore
