--- bots/balfagore/balfagore_main.lua
@@ -1,1662 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml">
-	<head>
-		<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-		<title>[Lua] ----------------------------------------------  -- _____         --|_   _|  --   - Pastebin.com</title>
-		<link rel="shortcut icon" href="/favicon.ico" />
-<link href="/cache/css/lua.css" rel="stylesheet" type="text/css" />		<link href="/i/main_v4.css" rel="stylesheet" type="text/css" />
-				
-		<script type="text/javascript" src="/js/jquery.js"></script>
-		<script type="text/javascript" src="/js/main_v1.js"></script>
-		<meta property="fb:app_id" content="231493360234820" />
-		<meta property="og:title" content="[Lua] ----------------------------------------------  -- _____         --|_   _|  --   - Pastebin.com" />
-		<meta property="og:type" content="article" />
-	    <meta property="og:url" content="http://pastebin.com/w7fj3tgY" />
-	    <meta property="og:image" content="http://pastebin.com/i/fb2.jpg" />
-	    <meta property="og:site_name" content="Pastebin" />
-		<meta name="google-site-verification" content="jkUAIOE8owUXu8UXIhRLB9oHJsWBfOgJbZzncqHoF4A" />
-		<link rel="canonical" href="http://pastebin.com/w7fj3tgY" />
-				<!--[if SafMob]>
-			<style>body{-webkit-text-size-adjust:none;}</style>
-		<![endif]-->
-		<script type="text/javascript">
-		  var _gaq = _gaq || [];
-		  _gaq.push(['_setAccount', 'UA-58643-34']);
-		  _gaq.push(['_trackPageview']);
-			setTimeout("_gaq.push(['_trackEvent', '15_seconds', 'read'])", 15000);
-		  (function() {
-		    var ga = document.createElement('script'); ga.type = 'text/javascript'; ga.async = true;
-		    ga.src = ('https:' == document.location.protocol ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js';
-		    var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(ga, s);
-		  })();
-		</script>
-	</head>
-	<body>
-	<div id="super_frame">
-		<div id="logo" onclick="location.href='/'" title="Create New Paste"></div>
-		<div id="header">
-			<div id="header_top">
-				<span class="span_left more">PASTEBIN</span><span class="span_left less"> &nbsp;|&nbsp; #1 paste tool since 2002</span><span class="min_max_span narrow_it" title="Change layout width"></span><span class="min_max_span wide_it" style="display:none" title="Change layout width"></span>				<ul class="top_menu">
-					<li class="no_border_li"><a href="/" accesskey="n">create new paste</a></li><li><a href="/tools">tools</a></li><li><a href="/api">api</a></li><li><a href="/archive">archive</a></li><li><a href="/faq">faq</a></li>
-				</ul>
-			</div>
-			<div id="header_middle">
-				<span class="span_left big"><a href="/">PASTEBIN</a></span> 
-				<a href="http://twitter.com/pastebin" target="_blank"><img src="/i/t.gif" alt="" class="i_tf" width="122" height="20" border="0" /></a>	<iframe src="//www.facebook.com/plugins/like.php?href=http%3A%2F%2Fwww.facebook.com%2Fpastebin&amp;send=false&amp;layout=button_count&amp;width=100&amp;show_faces=false&amp;action=like&amp;colorscheme=light&amp;font=segoe+ui&amp;height=21&amp;appId=231493360234820" scrolling="no" frameborder="0" style="border:none; overflow:hidden; width:100px; height:21px;margin:13px 0 0 10px;vertical-align:top" allowTransparency="true"></iframe>
-				<div id="header_middle_search">
-					<form class="search_form" name="search_form" method="get" action="/search" id="cse-search-box">
-					    <input type="hidden" name="cx" value="partner-pub-7089230323117142:2864958357" />
-					    <input type="hidden" name="cof" value="FORID:10" />
-					    <input type="hidden" name="ie" value="UTF-8" />
-						<input class="search_input" type="text" name="q" size="5" value="" placeholder="search..." x-webkit-speech/><input name="sa" class="search_button" src="/i/t.gif" alt="Search..." type="image" value="Search" />
-					</form>
-				</div>					
-			</div>
-			<div id="header_bottom">
-				<div class="div_top_menu">
-					<img src="/i/t.gif" class="i_n" width="16" height="16" alt="" border="0" /> <a href="/">create new paste</a> 
-					&nbsp;&nbsp;&nbsp; <img src="/i/t.gif" class="i_t" width="16" height="16" alt="" border="0" /> <a href="/trends">trending pastes</a>
-				</div>
-				<ul class="top_menu">
-					<li class="no_border_li"><a href="/signup">sign up</a></li><li><a href="/login">login</a></li><li><a href="/alerts">my alerts</a></li><li><a href="/settings">my settings</a></li><li><a href="/profile">my profile</a></li>				</ul>		
-			</div>			
-		</div>
-
-			<div class="frame_spacer"><!-- --></div>
-			<div style="height:35px;line-height:35px;font-size:0.85em;"><span style="float:right;text-align:right;">Don't like ads? <a href="/pro">PRO users</a> don't see any ads ;-)</span></div>		<div class="frame_spacer"><!-- --></div>
-		<div id="monster_frame">
-			<div id="content_frame">
-				<div id="content_right">
-										<div class="content_right_menu">
-									<div class="content_right_title"><a href="/archive">Public Pastes</a></div>	
-									<div id="menu_2">
-										<ul class="right_menu"><li><a href="/MRegk5sK">idgaf lol</a><span>11 sec ago</span></li><li><a href="/x0NEGT3E">Untitled</a><span>3 sec ago</span></li><li><a href="/tf8Wg1sq">New Signature</a><span>9 sec ago</span></li><li><a href="/jd6iRHun">Untitled</a><span>7 sec ago</span></li><li><a href="/QuJHiBV9">Untitled</a><span>9 sec ago</span></li><li><a href="/z9ZQf7T4">Tanoshi Aiura 11 720p mkv</a><span>9 sec ago</span></li><li><a href="/fir6AL3Q">Untitled</a><span>9 sec ago</span></li><li><a href="/hu52CQcn">Untitled</a><span>12 sec ago</span></li></ul></div></div>										
-			<div style="padding: 0; width:300px;height:260px;clear:left;">
-				<iframe src="/adserver/300x250_tribal_safe.html" width="300" height="250" scrolling="no" frameborder="0" name="ad3"></iframe>
-			</div>				</div>
-				<div id="content_left">
-	<div class="paste_box_frame">
-		<div class="tweet">
-			<div onclick="facebookpopup('/w7fj3tgY'); return false;" id="b_facebook2"><span id="b_facebook"></span></div>
-			<div onclick="twitpopup('/w7fj3tgY'); return false;" id="b_twitter2"><span id="b_twitter"></span></div>
-		</div>
-		<div class="paste_box_icon">
-			<img src="/i/t.gif" class="i_gb" alt="Guest" border="0" />	
-		</div>
-		<div class="paste_box_info">
-			<div class="paste_box_line1" title="Untitled"><img src="/i/t.gif"  class="i_p0" width="16" height="16" title="Public paste, everybody can see this paste." alt=""  border="0" /><h1>Untitled</h1> </div>
-			<div class="paste_box_line2">By: a guest  on <span title="Wednesday 22nd of May 2013 05:36:40 AM CDT" style="cursor:help">May 22nd, 2013</span> &nbsp;|&nbsp; syntax: <a href="/archive/lua">Lua</a> &nbsp;|&nbsp; size: 33.66 KB &nbsp;|&nbsp; hits: 33 &nbsp;|&nbsp; expires: Never</div>
-			<div class="paste_box_line3"><a href="/download.php?i=w7fj3tgY" rel="nofollow">download</a> &nbsp;|&nbsp; <a href="/raw.php?i=w7fj3tgY" target="_blank" rel="nofollow">raw</a> &nbsp;|&nbsp; <a href="/embed.php?i=w7fj3tgY" rel="nofollow">embed</a> &nbsp;|&nbsp; <a href="/report.php?i=w7fj3tgY" rel="nofollow">report abuse</a> &nbsp;|&nbsp; <a href="/print.php?i=w7fj3tgY" rel="nofollow">print</a></div>
-		</div>
-	</div>
-	
-			<div class="banner_728">
-				<iframe src="/adserver/728x90_tribal_safe.html" width="728" height="90" scrolling="no" frameborder="0" name="ad2"></iframe>
-			</div>
-			<div class="layout_clear"></div><div id="code_frame2">
-	<div id="code_frame">
-		<div id="code_buttons">
-		
-		<a href="javascript:togglev();" title="Show/Hide line numbers"><img src="/i/t.gif" border="0" alt="" class="i16 line" /></a> 
-		<a href="javascript:togglew('lua');" title="Toggle text wrapping"><img src="/i/t.gif" border="0" alt="" class="i16 wrap" /></a> 
-		<a href="#" class="copyme" onclick="selectText('selectable');showdiv('copied');" title="Copy text to clipboard"><img src="/i/t.gif" border="0" alt="" class="i16 clipboard" /></a> <span id="copied">Text below is selected. Please press Ctrl+C to copy to your clipboard. (&#8984;+C on Mac)</span>
-		</div>
-		<div id="selectable">		
-		<div class="lua"><ol><li class="li1"><div class="de1"><span class="co1">----------------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">-- _____ &nbsp; &nbsp; &nbsp; </span></div></li>
-<li class="li1"><div class="de1"><span class="co1">--|_ &nbsp; _|</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">-- &nbsp;| | &nbsp; _ __ &nbsp; __ _ &nbsp;_ __ &nbsp; &nbsp;__ _ &nbsp; __ _ &nbsp; ___ &nbsp; _ __ &nbsp; ___</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">-- &nbsp;| | &nbsp;| '__| / _` || '_ \ &nbsp;/ _` | / _` | / _ \ | '__| / _ \</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">-- &nbsp;| | &nbsp;| | &nbsp; | (_| || |_) || (_| || (_| || (_) || | &nbsp; | &nbsp;__/</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">-- &nbsp;\_/ &nbsp;|_| &nbsp; &nbsp;\__,_|| .__/ &nbsp;\__,_| \__, | \___/ |_| &nbsp; &nbsp;\___|</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">-- &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;| | &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; __/ |</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">-- &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;|_| &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;|___/ &nbsp; &nbsp; version 0.8.3</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">--</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">------------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">-- &nbsp; &nbsp; &nbsp; &nbsp; Created by: kairus101 &nbsp; &nbsp; &nbsp; &nbsp;--</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">------------------------------------------</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1"><span class="co1">------------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">-- &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;Bot Initialization &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;--</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">------------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">local</span> <span class="kw3">_G</span> <span class="sy0">=</span> <span class="kw3">getfenv</span><span class="br0">&#40;</span><span class="nu0">0</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1"><span class="kw1">local</span> object <span class="sy0">=</span> <span class="kw3">_G</span><span class="sy0">.</span>object</div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>myName <span class="sy0">=</span> object<span class="sy0">:</span>GetName<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>bRunLogic<span class="sy0">,</span>object<span class="sy0">.</span>bRunBehaviors<span class="sy0">,</span>object<span class="sy0">.</span>bUpdates<span class="sy0">,</span>object<span class="sy0">.</span>bUseShop<span class="sy0">,</span>object<span class="sy0">.</span>bRunCommands<span class="sy0">,</span>object<span class="sy0">.</span>bMoveCommands<span class="sy0">,</span>object<span class="sy0">.</span>bAttackCommands<span class="sy0">,</span>object<span class="sy0">.</span>bAbilityCommands<span class="sy0">,</span>object<span class="sy0">.</span>bOtherCommands<span class="sy0">=</span><span class="kw4">true</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>logger <span class="sy0">=</span> <span class="br0">&#123;</span><span class="br0">&#125;</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>bReportBehavior<span class="sy0">,</span>object<span class="sy0">.</span>bDebugUtility<span class="sy0">,</span>object<span class="sy0">.</span>logger<span class="sy0">.</span>bWriteLog<span class="sy0">,</span>object<span class="sy0">.</span>logger<span class="sy0">.</span>bVerboseLog<span class="sy0">=</span><span class="kw4">false</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>core &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="sy0">=</span> <span class="br0">&#123;</span><span class="br0">&#125;</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>eventsLib &nbsp; &nbsp; &nbsp; &nbsp;<span class="sy0">=</span> <span class="br0">&#123;</span><span class="br0">&#125;</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>metadata &nbsp; &nbsp; &nbsp; &nbsp; <span class="sy0">=</span> <span class="br0">&#123;</span><span class="br0">&#125;</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>behaviorLib &nbsp; &nbsp; &nbsp;<span class="sy0">=</span> <span class="br0">&#123;</span><span class="br0">&#125;</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>skills &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="sy0">=</span> <span class="br0">&#123;</span><span class="br0">&#125;</span></div></li>
-<li class="li1"><div class="de1">runfile <span class="st0">&quot;bots/core.lua&quot;</span></div></li>
-<li class="li2"><div class="de2">runfile <span class="st0">&quot;bots/botbraincore.lua&quot;</span></div></li>
-<li class="li1"><div class="de1">runfile <span class="st0">&quot;bots/eventslib.lua&quot;</span></div></li>
-<li class="li2"><div class="de2">runfile <span class="st0">&quot;bots/metadata.lua&quot;</span></div></li>
-<li class="li1"><div class="de1">runfile <span class="st0">&quot;bots/behaviorlib.lua&quot;</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">local</span> core<span class="sy0">,</span> eventsLib<span class="sy0">,</span> behaviorLib<span class="sy0">,</span> metadata<span class="sy0">,</span> skills <span class="sy0">=</span> object<span class="sy0">.</span>core<span class="sy0">,</span> object<span class="sy0">.</span>eventsLib<span class="sy0">,</span> object<span class="sy0">.</span>behaviorLib<span class="sy0">,</span> object<span class="sy0">.</span>metadata<span class="sy0">,</span> object<span class="sy0">.</span>skills</div></li>
-<li class="li1"><div class="de1"><span class="kw1">local</span> <span class="kw3">print</span><span class="sy0">,</span> <span class="kw3">ipairs</span><span class="sy0">,</span> <span class="kw3">pairs</span><span class="sy0">,</span> <span class="kw3">string</span><span class="sy0">,</span> <span class="kw3">table</span><span class="sy0">,</span> <span class="kw3">next</span><span class="sy0">,</span> <span class="kw3">type</span><span class="sy0">,</span> <span class="kw3">tinsert</span><span class="sy0">,</span> <span class="kw3">tremove</span><span class="sy0">,</span> tsort<span class="sy0">,</span> <span class="kw3">format</span><span class="sy0">,</span> <span class="kw3">tostring</span><span class="sy0">,</span> <span class="kw3">tonumber</span><span class="sy0">,</span> <span class="kw3">strfind</span><span class="sy0">,</span> <span class="kw3">strsub</span> &nbsp; &nbsp; <span class="sy0">=</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">print</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">ipairs</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">pairs</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">string</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">table</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">next</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">type</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">table.insert</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">table.remove</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">table.sort</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">string.format</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">tostring</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">tonumber</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">string.find</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">string.sub</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">local</span> <span class="kw3">ceil</span><span class="sy0">,</span> <span class="kw3">floor</span><span class="sy0">,</span> pi<span class="sy0">,</span> <span class="kw3">tan</span><span class="sy0">,</span> <span class="kw3">atan</span><span class="sy0">,</span> <span class="kw3">atan2</span><span class="sy0">,</span> <span class="kw3">abs</span><span class="sy0">,</span> <span class="kw3">cos</span><span class="sy0">,</span> <span class="kw3">sin</span><span class="sy0">,</span> <span class="kw3">acos</span><span class="sy0">,</span> <span class="kw3">max</span><span class="sy0">,</span> <span class="kw3">random</span><span class="sy0">,</span> <span class="kw3">sqrt</span> <span class="sy0">=</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">math.ceil</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">math.floor</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">math.pi</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">math.tan</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">math.atan</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">math.atan2</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">math.abs</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">math.cos</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">math.sin</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">math.acos</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">math.max</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">math.random</span><span class="sy0">,</span> <span class="kw3">_G</span><span class="sy0">.</span><span class="kw3">math.sqrt</span></div></li>
-<li class="li1"><div class="de1"><span class="kw1">local</span> BotEcho<span class="sy0">,</span> VerboseLog<span class="sy0">,</span> BotLog <span class="sy0">=</span> core<span class="sy0">.</span>BotEcho<span class="sy0">,</span> core<span class="sy0">.</span>VerboseLog<span class="sy0">,</span> core<span class="sy0">.</span>BotLog</div></li>
-<li class="li2"><div class="de2"><span class="kw1">local</span> Clamp <span class="sy0">=</span> core<span class="sy0">.</span>Clamp</div></li>
-<li class="li1"><div class="de1">BotEcho<span class="br0">&#40;</span><span class="st0">'loading balfagore_main...'</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1"><span class="co1">---------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">-- &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;Constants &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;--</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">---------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">--Balphagore</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>heroName <span class="sy0">=</span> <span class="st0">'Hero_Bephelgor'</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1"><span class="co1">-- Item buy order. internal names</span></div></li>
-<li class="li2"><div class="de2">behaviorLib<span class="sy0">.</span>StartingItems <span class="sy0">=</span> <span class="br0">&#123;</span><span class="st0">&quot;Item_LoggersHatchet&quot;</span><span class="sy0">,</span> <span class="st0">&quot;Item_GuardianRing&quot;</span><span class="sy0">,</span> <span class="st0">&quot;2 Item_RunesOfTheBlight&quot;</span><span class="br0">&#125;</span></div></li>
-<li class="li1"><div class="de1">behaviorLib<span class="sy0">.</span>LaneItems <span class="sy0">=</span> <span class="br0">&#123;</span><span class="st0">&quot;Item_ManaRegen3&quot;</span><span class="sy0">,</span><span class="st0">&quot;Item_Marchers&quot;</span><span class="sy0">,</span><span class="st0">&quot;Item_MysticVestments&quot;</span><span class="sy0">,</span><span class="st0">&quot;Item_EnhancedMarchers&quot;</span><span class="br0">&#125;</span></div></li>
-<li class="li2"><div class="de2">behaviorLib<span class="sy0">.</span>MidItems <span class="sy0">=</span> <span class="br0">&#123;</span><span class="st0">&quot;Item_Energizer&quot;</span><span class="sy0">,</span><span class="st0">&quot;Item_Pierce 3&quot;</span><span class="sy0">,</span> <span class="st0">&quot;Item_SolsBulwark&quot;</span><span class="sy0">,</span> <span class="st0">&quot;Item_DaemonicBreastplate&quot;</span><span class="br0">&#125;</span></div></li>
-<li class="li1"><div class="de1">behaviorLib<span class="sy0">.</span>LateItems <span class="sy0">=</span> <span class="br0">&#123;</span><span class="st0">&quot;Item_BehemothsHeart&quot;</span><span class="sy0">,</span> <span class="st0">'Item_Damage9'</span><span class="br0">&#125;</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1"><span class="co1">-- Skillbuild. 0 is Regurgitate, 1 is Demonic Pathogen, 2 is Corpse Conversion, 3 is Hell on Newerth, 4 is Attributes</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>tSkills <span class="sy0">=</span> <span class="br0">&#123;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="nu0">2</span><span class="sy0">,</span> <span class="nu0">0</span><span class="sy0">,</span> <span class="nu0">2</span><span class="sy0">,</span> <span class="nu0">0</span><span class="sy0">,</span> <span class="nu0">2</span><span class="sy0">,</span> <span class="nu0">0</span><span class="sy0">,</span> &nbsp; &nbsp; &nbsp; <span class="co1">-- 1-6</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="nu0">2</span><span class="sy0">,</span> <span class="nu0">0</span><span class="sy0">,</span> <span class="nu0">3</span><span class="sy0">,</span> <span class="nu0">4</span><span class="sy0">,</span> <span class="nu0">3</span><span class="sy0">,</span> &nbsp;&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- 7-11</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="nu0">4</span><span class="sy0">,</span> <span class="nu0">4</span><span class="sy0">,</span> <span class="nu0">4</span><span class="sy0">,</span> <span class="nu0">4</span><span class="sy0">,</span> <span class="nu0">3</span><span class="sy0">,</span>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- 12-16</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="nu0">4</span><span class="sy0">,</span> <span class="nu0">4</span><span class="sy0">,</span> <span class="nu0">4</span><span class="sy0">,</span> <span class="nu0">4</span><span class="sy0">,</span> <span class="nu0">4</span><span class="sy0">,</span> <span class="nu0">1</span><span class="sy0">,</span> <span class="nu0">1</span><span class="sy0">,</span> <span class="nu0">1</span><span class="sy0">,</span> <span class="nu0">1</span><span class="sy0">,</span>&nbsp; &nbsp; &nbsp; <span class="co1">--17-25</span></div></li>
-<li class="li1"><div class="de1"><span class="br0">&#125;</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1"><span class="co1">-- Bonus agression points if a skill/item is available for use</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>nHellUp <span class="sy0">=</span> <span class="nu0">5</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>nRegurgitateUp <span class="sy0">=</span> <span class="nu0">10</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>nMinionsCloseUp <span class="sy0">=</span> <span class="nu0">8</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>nTrappingUp <span class="sy0">=</span> <span class="nu0">10</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>nEnergizerUp <span class="sy0">=</span> <span class="nu0">8</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>nCanAttackUp <span class="sy0">=</span> <span class="nu0">10</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1"><span class="co1">-- Bonus agression points that are applied to the bot upon successfully using a skill/item</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>nRegurgitateUse <span class="sy0">=</span> <span class="nu0">10</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>nHellUse <span class="sy0">=</span> <span class="nu0">8</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>nEnergizerUse <span class="sy0">=</span> <span class="nu0">8</span></div></li>
-<li class="li1"><div class="de1">&nbsp;</div></li>
-<li class="li2"><div class="de2"><span class="co1">-- Thresholds of aggression the bot must reach to use these abilities</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>nRegurgitateThreshold <span class="sy0">=</span> <span class="nu0">26</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>tHellOnNewerthThreshold <span class="sy0">=</span> <span class="br0">&#123;</span><span class="nu0">70</span><span class="sy0">,</span><span class="nu0">60</span><span class="sy0">,</span><span class="nu0">50</span><span class="br0">&#125;</span> <span class="co1">--depending on charges</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>nEnergizerThreshold <span class="sy0">=</span> <span class="nu0">30</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1"><span class="co1">--time variables</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>nTimeBarfed<span class="sy0">=</span><span class="nu0">0</span> <span class="co1">--used to get corpses at well</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>nTimeEnergizered<span class="sy0">=</span><span class="nu0">0</span> <span class="co1">--used for minions to remember when to trap again</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1"><span class="co1">--table variables (need to keep track on minions and corpses)</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>tMinions<span class="sy0">=</span><span class="br0">&#123;</span><span class="br0">&#125;</span> <span class="co1">--hold minions</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>tCorpseTable<span class="sy0">=</span><span class="br0">&#123;</span><span class="br0">&#125;</span><span class="co1">--position,time</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>tOldLocalCreeps<span class="sy0">=</span><span class="br0">&#123;</span><span class="br0">&#125;</span><span class="co1">--for checking if units are dead for corpses.</span></div></li>
-<li class="li1"><div class="de1">&nbsp;</div></li>
-<li class="li2"><div class="de2"><span class="co1">--item variables</span></div></li>
-<li class="li1"><div class="de1">core<span class="sy0">.</span>itemEnergizer<span class="sy0">=</span><span class="kw4">nil</span> <span class="co1">--energizer</span></div></li>
-<li class="li2"><div class="de2">core<span class="sy0">.</span>itemBols<span class="sy0">=</span><span class="kw4">nil</span> <span class="co1">--sol's bolwork</span></div></li>
-<li class="li1"><div class="de1">&nbsp;</div></li>
-<li class="li2"><div class="de2"><span class="co1">--other variables</span></div></li>
-<li class="li1"><div class="de1">core<span class="sy0">.</span>distSqTolerance<span class="sy0">=</span><span class="nu0">5</span><span class="sy0">*</span><span class="nu0">5</span></div></li>
-<li class="li2"><div class="de2">behaviorLib<span class="sy0">.</span>nCreepPushbackMul <span class="sy0">=</span> <span class="nu0">0.4</span> <span class="co1">--0.5</span></div></li>
-<li class="li1"><div class="de1">behaviorLib<span class="sy0">.</span>nTargetPositioningMul <span class="sy0">=</span> <span class="nu0">0.4</span> <span class="co1">--0.6</span></div></li>
-<li class="li2"><div class="de2">behaviorLib<span class="sy0">.</span>nTargetCriticalPositioningMul <span class="sy0">=</span> <span class="nu0">1</span> <span class="co1">--2</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>nRealDistance<span class="sy0">=</span><span class="nu0">0</span> <span class="co1">--average distance minions are from target.</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>nMinionsClose<span class="sy0">=</span><span class="nu0">0</span> <span class="co1">--this is for use with energizer</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>bUnitStill<span class="sy0">=</span><span class="kw4">false</span> <span class="co1">--this adds extra harass utility if hero is trapped</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>nMinionSkip<span class="sy0">=</span><span class="nu0">5</span> <span class="co1">--control minions every object.nMinionSkip frames.</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>nCurSkip<span class="sy0">=</span><span class="nu0">0</span> <span class="co1">--current frame. This is for frame skip.</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1"><span class="co1">--difficulty variables</span></div></li>
-<li class="li2"><div class="de2"><span class="coMULTI">--[[</span></div></li>
-<li class="li1"><div class="de1"><span class="coMULTI">core.nEASY_DIFFICULTY &nbsp; = 1</span></div></li>
-<li class="li2"><div class="de2"><span class="coMULTI">core.nMEDIUM_DIFFICULTY = 2</span></div></li>
-<li class="li1"><div class="de1"><span class="coMULTI">core.nHARD_DIFFICULTY &nbsp; = 3</span></div></li>
-<li class="li2"><div class="de2"><span class="coMULTI">]]</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>trapEffectiveness<span class="sy0">=</span><span class="br0">&#123;</span><span class="nu0">95</span><span class="sy0">*</span><span class="nu0">95</span><span class="sy0">,</span><span class="nu0">75</span><span class="sy0">*</span><span class="nu0">75</span><span class="sy0">,</span><span class="nu0">55</span><span class="sy0">*</span><span class="nu0">55</span><span class="br0">&#125;</span><span class="co1">--Lower is better.</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>trapCycleSkip<span class="sy0">=</span><span class="br0">&#123;</span><span class="nu0">7</span><span class="sy0">,</span><span class="nu0">4</span><span class="sy0">,</span><span class="nu0">1</span><span class="br0">&#125;</span><span class="co1">--Lower is better.</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>trapCircleRadius<span class="sy0">=</span><span class="br0">&#123;</span><span class="nu0">550</span><span class="sy0">,</span><span class="nu0">450</span><span class="sy0">,</span><span class="nu0">350</span><span class="br0">&#125;</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1"><span class="co1">------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">-- &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;Skills &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;--</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">function</span> object<span class="sy0">:</span>SkillBuild<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> unitSelf <span class="sy0">=</span> self<span class="sy0">.</span>core<span class="sy0">.</span>unitSelf</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> &nbsp;skills<span class="sy0">.</span>abilRegurgitate <span class="sy0">==</span> <span class="kw4">nil</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; skills<span class="sy0">.</span>abilRegurgitate <span class="sy0">=</span> unitSelf<span class="sy0">:</span>GetAbility<span class="br0">&#40;</span><span class="nu0">0</span><span class="br0">&#41;</span><span class="co1">--Barf</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; skills<span class="sy0">.</span>abilDemonicPathogen <span class="sy0">=</span> unitSelf<span class="sy0">:</span>GetAbility<span class="br0">&#40;</span><span class="nu0">1</span><span class="br0">&#41;</span><span class="co1">--Useless silence</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; skills<span class="sy0">.</span>abilSpawnMinions <span class="sy0">=</span> unitSelf<span class="sy0">:</span>GetAbility<span class="br0">&#40;</span><span class="nu0">2</span><span class="br0">&#41;</span><span class="co1">--Corpse pickup/Minions</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; skills<span class="sy0">.</span>abilHellOnNewerth <span class="sy0">=</span> unitSelf<span class="sy0">:</span>GetAbility<span class="br0">&#40;</span><span class="nu0">3</span><span class="br0">&#41;</span><span class="co1">--Ultimate</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; skills<span class="sy0">.</span>abilAttributeBoost <span class="sy0">=</span> unitSelf<span class="sy0">:</span>GetAbility<span class="br0">&#40;</span><span class="nu0">4</span><span class="br0">&#41;</span><span class="co1">--Stats</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> unitSelf<span class="sy0">:</span>GetAbilityPointsAvailable<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="sy0">&lt;=</span> <span class="nu0">0</span> <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nlev <span class="sy0">=</span> unitSelf<span class="sy0">:</span>GetLevel<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nlevpts <span class="sy0">=</span> unitSelf<span class="sy0">:</span>GetAbilityPointsAvailable<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">for</span> i <span class="sy0">=</span> nlev<span class="sy0">,</span> nlev<span class="sy0">+</span>nlevpts <span class="kw1">do</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; unitSelf<span class="sy0">:</span>GetAbility<span class="br0">&#40;</span> object<span class="sy0">.</span>tSkills<span class="br0">&#91;</span>i<span class="br0">&#93;</span> <span class="br0">&#41;</span><span class="sy0">:</span>LevelUp<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1"><span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1"><span class="co1">------------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">-- &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;FindItems Override &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;--</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">------------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">local</span> <span class="kw1">function</span> FindItemsOverride<span class="br0">&#40;</span>botBrain<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> bUpdated <span class="sy0">=</span> object<span class="sy0">.</span>FindItemsOld<span class="br0">&#40;</span>botBrain<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> core<span class="sy0">.</span>itemEnergizer <span class="sy0">~=</span> <span class="kw4">nil</span> <span class="kw2">and</span> <span class="kw2">not</span> core<span class="sy0">.</span>itemEnergizer<span class="sy0">:</span>IsValid<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; core<span class="sy0">.</span>itemEnergizer <span class="sy0">=</span> <span class="kw4">nil</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> core<span class="sy0">.</span>itemGhostMarchers <span class="sy0">~=</span> <span class="kw4">nil</span> <span class="kw2">and</span> <span class="kw2">not</span> core<span class="sy0">.</span>itemGhostMarchers<span class="sy0">:</span>IsValid<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; core<span class="sy0">.</span>itemGhostMarchers <span class="sy0">=</span> <span class="kw4">nil</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp;</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> bUpdated <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> core<span class="sy0">.</span>itemEnergizer <span class="kw2">and</span> core<span class="sy0">.</span>itemGhostMarchers <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> inventory <span class="sy0">=</span> core<span class="sy0">.</span>unitSelf<span class="sy0">:</span>GetInventory<span class="br0">&#40;</span><span class="kw4">false</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">for</span> slot <span class="sy0">=</span> <span class="nu0">1</span><span class="sy0">,</span> <span class="nu0">6</span> <span class="kw1">do</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> curItem <span class="sy0">=</span> inventory<span class="br0">&#91;</span>slot<span class="br0">&#93;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> curItem <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> core<span class="sy0">.</span>itemEnergizer <span class="sy0">==</span> <span class="kw4">nil</span> <span class="kw2">and</span> curItem<span class="sy0">:</span>GetName<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="sy0">==</span> <span class="st0">&quot;Item_Energizer&quot;</span> <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; core<span class="sy0">.</span>itemEnergizer <span class="sy0">=</span> core<span class="sy0">.</span>WrapInTable<span class="br0">&#40;</span>curItem<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">elseif</span> core<span class="sy0">.</span>itemGhostMarchers <span class="sy0">==</span> <span class="kw4">nil</span> <span class="kw2">and</span> curItem<span class="sy0">:</span>GetName<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="sy0">==</span> <span class="st0">&quot;Item_EnhancedMarchers&quot;</span> <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; core<span class="sy0">.</span>itemGhostMarchers <span class="sy0">=</span> core<span class="sy0">.</span>WrapInTable<span class="br0">&#40;</span>curItem<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">elseif</span> core<span class="sy0">.</span>itemBols <span class="sy0">==</span> <span class="kw4">nil</span> <span class="kw2">and</span> curItem<span class="sy0">:</span>GetName<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="sy0">==</span> <span class="st0">&quot;Item_SolsBulwark&quot;</span> <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; core<span class="sy0">.</span>itemBols <span class="sy0">=</span> core<span class="sy0">.</span>WrapInTable<span class="br0">&#40;</span>curItem<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; core<span class="sy0">.</span>OrderItemClamp<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> unitSelf<span class="sy0">,</span> core<span class="sy0">.</span>itemBols<span class="br0">&#41;</span> <span class="co1">--turn to negative bols</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>FindItemsOld<span class="sy0">=</span>core<span class="sy0">.</span>FindItems</div></li>
-<li class="li2"><div class="de2">core<span class="sy0">.</span>FindItems <span class="sy0">=</span> FindItemsOverride</div></li>
-<li class="li1"><div class="de1">&nbsp;</div></li>
-<li class="li2"><div class="de2"><span class="co1">----------------------------------------------------</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">-- &nbsp; &nbsp; &nbsp; &nbsp;Minion Trapping Helper Functions &nbsp; &nbsp; &nbsp; &nbsp;--</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">----------------------------------------------------</span></div></li>
-<li class="li1"><div class="de1"><span class="kw1">local</span> <span class="kw1">function</span> GetHeroByID<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> ID<span class="br0">&#41;</span> <span class="co1">--refreshes the targetted unit.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">for</span> id<span class="sy0">,</span> unitHero <span class="kw2">in</span> <span class="kw3">pairs</span><span class="br0">&#40;</span>HoN<span class="sy0">.</span>GetHeroes<span class="br0">&#40;</span>core<span class="sy0">.</span>enemyTeam<span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="kw1">do</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> unitHero <span class="sy0">~=</span> <span class="kw4">nil</span> <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> core<span class="sy0">.</span>CanSeeUnit<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> unitHero<span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> unitHero<span class="sy0">:</span>GetUniqueID<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">==</span>ID <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> unitHero</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> <span class="kw4">nil</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1"><span class="kw1">local</span> <span class="kw1">function</span> positionOffset<span class="br0">&#40;</span>pos<span class="sy0">,</span> angle<span class="sy0">,</span> distance<span class="br0">&#41;</span> <span class="co1">--this is used by minions to form a ring around people.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; tmp <span class="sy0">=</span> Vector3<span class="sy0">.</span>Create<span class="br0">&#40;</span><span class="kw3">cos</span><span class="br0">&#40;</span>angle<span class="br0">&#41;</span><span class="sy0">*</span>distance<span class="sy0">,</span><span class="kw3">sin</span><span class="br0">&#40;</span>angle<span class="br0">&#41;</span><span class="sy0">*</span>distance<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> tmp<span class="sy0">+</span>pos</div></li>
-<li class="li2"><div class="de2"><span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1"><span class="kw1">local</span> <span class="kw1">function</span> vectorLength<span class="br0">&#40;</span>vec<span class="br0">&#41;</span> <span class="co1">--easy way to get the distance between two points. Much more efficient than pythagoras' theorem - thanks to Anakonda.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> x <span class="sy0">=</span> vec<span class="sy0">.</span>x</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> y <span class="sy0">=</span> vec<span class="sy0">.</span>y</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> x<span class="sy0">==</span><span class="nu0">0</span> <span class="kw1">then</span> <span class="kw1">return</span> y <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> y<span class="sy0">==</span><span class="nu0">0</span> <span class="kw1">then</span> <span class="kw1">return</span> x <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> angle <span class="sy0">=</span> <span class="kw3">atan</span><span class="br0">&#40;</span>y<span class="sy0">/</span>x<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> y<span class="sy0">/</span><span class="kw3">sin</span><span class="br0">&#40;</span>angle<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp;</div></li>
-<li class="li2"><div class="de2"><span class="co1">----------------------------------------</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">-- &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;OnThink Override &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;--</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">----------------------------------------</span></div></li>
-<li class="li1"><div class="de1"><span class="kw1">function</span> object<span class="sy0">:</span>onthinkOverride<span class="br0">&#40;</span>tGameVariables<span class="br0">&#41;</span> <span class="co1">--This is run, even while dead. Every frame.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; self<span class="sy0">:</span>onthinkOld<span class="br0">&#40;</span>tGameVariables<span class="br0">&#41;</span><span class="co1">--don't distrupt old think, run it.</span></div></li>
-<li class="li1"><div class="de1">&nbsp;</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> unitSelf <span class="sy0">=</span> core<span class="sy0">.</span>unitSelf</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> vecSelfPos <span class="sy0">=</span> unitSelf<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nNumMinions<span class="sy0">=</span><span class="kw3">table.getn</span><span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> bDebugLines <span class="sy0">=</span> <span class="kw4">false</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">---------------------------------</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- &nbsp; &nbsp; &nbsp;Register Minions &nbsp; &nbsp; &nbsp; --</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">---------------------------------</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>nNumMinions<span class="sy0">==</span><span class="nu0">0</span> <span class="kw2">and</span> unitSelf<span class="sy0">:</span>IsAlive<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> core<span class="sy0">.</span>localUnits<span class="sy0">~=</span><span class="kw4">nil</span><span class="br0">&#41;</span><span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; object<span class="sy0">.</span>tMinions<span class="sy0">=</span><span class="br0">&#123;</span><span class="br0">&#125;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; i<span class="sy0">=</span><span class="nu0">1</span><span class="co1">--keep track of array index</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">for</span> key<span class="sy0">,</span> unit <span class="kw2">in</span> <span class="kw3">pairs</span><span class="br0">&#40;</span>core<span class="sy0">.</span>localUnits<span class="br0">&#91;</span><span class="st0">&quot;AllyUnits&quot;</span><span class="br0">&#93;</span><span class="br0">&#41;</span> <span class="kw1">do</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>unit<span class="sy0">:</span>GetTypeName<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">==</span><span class="st0">&quot;Pet_Bephelgor_Ability2&quot;</span> <span class="kw2">and</span> unit<span class="sy0">:</span>GetOwnerPlayer<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">==</span>unitSelf<span class="sy0">:</span>GetOwnerPlayer<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">=</span>unit</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>bDebugLines<span class="br0">&#41;</span> <span class="kw1">then</span> core<span class="sy0">.</span>DrawDebugArrow<span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> vecSelfPos<span class="sy0">,</span> <span class="st0">'white'</span><span class="br0">&#41;</span> <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; i<span class="sy0">=</span>i<span class="sy0">+</span><span class="nu0">1</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-----------------------------------------------------------------------------------</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">---------------------------------</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- &nbsp; &nbsp; &nbsp; Control Minions &nbsp; &nbsp; &nbsp; --</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">---------------------------------</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- This part is the most complicated part of the program. The minions:</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- 1. Find a hero or building.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- 2. if target is a building, attack it.</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- 3. If it's a hero:</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- &nbsp; a. If target/minion is unit walking, just attack them.</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- &nbsp; b. If target is not trapped yet and we have &gt;3 minions, surround it,</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- &nbsp; &nbsp; &nbsp; &nbsp;and close in until they are trapped.</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- &nbsp; c. Attack trapped target.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- 4. If there is no target, attack-move along with the creep wave.</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-----------------------------------------------------------------------------------</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--this is a frame-skip. We don't need to run minions code every single frame.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; object<span class="sy0">.</span>nCurSkip<span class="sy0">=</span>object<span class="sy0">.</span>nCurSkip<span class="sy0">+</span><span class="nu0">1</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>object<span class="sy0">.</span>nCurSkip<span class="sy0">&gt;=</span>object<span class="sy0">.</span>nMinionSkip <span class="kw2">and</span> <span class="kw3">table.getn</span><span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#41;</span><span class="sy0">&gt;</span><span class="nu0">0</span><span class="br0">&#41;</span> <span class="kw1">then</span><span class="co1">--don't skip</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; object<span class="sy0">.</span>nCurSkip<span class="sy0">=</span><span class="nu0">0</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> botBrain<span class="sy0">=</span>self</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> bDebugLines <span class="sy0">=</span> <span class="kw4">false</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> vecWellPos<span class="sy0">=</span><span class="br0">&#40;</span>core<span class="sy0">.</span>enemyWell <span class="kw2">and</span> core<span class="sy0">.</span>enemyWell<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">---------------------------------</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- &nbsp; &nbsp;1.Minions Get Target &nbsp; &nbsp; --</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--------------------------------- &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> target<span class="sy0">=</span><span class="kw4">nil</span> <span class="co1">--this is what minions will use</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> heroTarget<span class="sy0">=</span>behaviorLib<span class="sy0">.</span>heroTarget</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> buildingTarget<span class="sy0">=</span>behaviorLib<span class="sy0">.</span>hitBuildingTarget</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>heroTarget <span class="kw2">and</span> heroTarget<span class="sy0">:</span>IsValid<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> heroTarget<span class="sy0">:</span>IsAlive<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> heroTarget<span class="sy0">:</span>GetHealth<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">&gt;</span><span class="nu0">0</span> <span class="kw2">and</span> core<span class="sy0">.</span>CanSeeUnit<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> heroTarget<span class="br0">&#41;</span> <span class="kw2">and</span> Vector3<span class="sy0">.</span>Distance2DSq<span class="br0">&#40;</span>heroTarget<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> vecSelfPos<span class="br0">&#41;</span><span class="sy0">&lt;</span><span class="nu0">2000</span><span class="sy0">*</span><span class="nu0">2000</span><span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--If we have a hero in mind</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; target<span class="sy0">=</span>GetHeroByID<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> heroTarget<span class="sy0">:</span>GetUniqueID<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">elseif</span> <span class="br0">&#40;</span>buildingTarget <span class="kw2">and</span> buildingTarget<span class="sy0">:</span>IsValid<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> buildingTarget<span class="sy0">:</span>IsAlive<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> core<span class="sy0">.</span>CanSeeUnit<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> buildingTarget<span class="br0">&#41;</span> <span class="kw2">and</span> Vector3<span class="sy0">.</span>Distance2DSq<span class="br0">&#40;</span>buildingTarget<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> vecSelfPos<span class="br0">&#41;</span><span class="sy0">&lt;</span><span class="nu0">2000</span><span class="sy0">*</span><span class="nu0">2000</span><span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--If we have a building in mind</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; target<span class="sy0">=</span>buildingTarget</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; object<span class="sy0">.</span>nRealDistance<span class="sy0">=</span><span class="kw4">nil</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; object<span class="sy0">.</span>nMinionsClose<span class="sy0">=</span><span class="nu0">0</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>target <span class="kw2">and</span> bDebugLines<span class="br0">&#41;</span> <span class="kw1">then</span> core<span class="sy0">.</span>DrawDebugArrow<span class="br0">&#40;</span>core<span class="sy0">.</span>unitSelf<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> target<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> <span class="st0">'white'</span><span class="br0">&#41;</span> <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">---------------------------------</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- &nbsp;2.Minions Building Attack &nbsp;--</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">---------------------------------</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>target <span class="kw2">and</span> target<span class="sy0">:</span>IsBuilding<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="kw1">then</span> <span class="co1">--kill buildings trapagore attacks, first and foremost.</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">for</span> i<span class="sy0">=</span><span class="nu0">1</span><span class="sy0">,</span>nNumMinions <span class="kw1">do</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span> <span class="kw2">and</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>IsValid<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>IsAlive<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; core<span class="sy0">.</span>OrderAttack<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">,</span> target<span class="sy0">,</span><span class="kw4">false</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; object<span class="sy0">.</span>nMinionSkip<span class="sy0">=</span><span class="nu0">10</span><span class="co1">--500ms till next cycle</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">---------------------------------</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- &nbsp; &nbsp;3.Minions Hero Attack &nbsp; &nbsp;--</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">---------------------------------</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">elseif</span> <span class="br0">&#40;</span>target <span class="kw2">and</span> nNumMinions<span class="sy0">&gt;</span><span class="nu0">2</span><span class="br0">&#41;</span><span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> target<span class="sy0">=</span>target</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> vecTargetPos<span class="sy0">=</span>target<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="co1">--+vTargetVel*0.5</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> angleToWell<span class="sy0">=</span> <span class="kw3">atan2</span><span class="br0">&#40;</span>vecWellPos<span class="sy0">.</span>y<span class="sy0">-</span>vecTargetPos<span class="sy0">.</span>y<span class="sy0">,</span>vecWellPos<span class="sy0">.</span>x<span class="sy0">-</span>vecTargetPos<span class="sy0">.</span>x<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>nNumMinions<span class="sy0">&gt;</span><span class="nu0">3</span><span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; object<span class="sy0">.</span>nMinionSkip<span class="sy0">=</span>object<span class="sy0">.</span>trapCycleSkip<span class="br0">&#91;</span>core<span class="sy0">.</span>nDifficulty<span class="br0">&#93;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> totalDistance<span class="sy0">=</span><span class="nu0">0</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">for</span> i<span class="sy0">=</span><span class="nu0">1</span><span class="sy0">,</span>nNumMinions <span class="kw1">do</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span> <span class="kw2">and</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>IsValid<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>IsAlive<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> target <span class="kw2">and</span> vecTargetPos<span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; vecMinionPos<span class="sy0">=</span>object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; totalDistance<span class="sy0">=</span>totalDistance<span class="sy0">+</span>vectorLength<span class="br0">&#40;</span>vecMinionPos<span class="sy0">-</span>vecTargetPos<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--local nAvgDistance=1.1*sqrt(tTotalDisplacement[1]*tTotalDisplacement[1]+tTotalDisplacement[2]*tTotalDisplacement[2])/nNumMinions</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nAvgDistance<span class="sy0">=</span>totalDistance<span class="sy0">/</span>nNumMinions</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; object<span class="sy0">.</span>nRealDistance<span class="sy0">=</span>nAvgDistance</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; distance<span class="sy0">=</span>nAvgDistance<span class="sy0">-</span><span class="nu0">30</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>distance<span class="sy0">&gt;</span>object<span class="sy0">.</span>trapCircleRadius<span class="br0">&#91;</span>core<span class="sy0">.</span>nDifficulty<span class="br0">&#93;</span><span class="br0">&#41;</span><span class="kw1">then</span> distance<span class="sy0">=</span>object<span class="sy0">.</span>trapCircleRadius<span class="br0">&#91;</span>core<span class="sy0">.</span>nDifficulty<span class="br0">&#93;</span> <span class="kw1">end</span> <span class="co1">--There is a cap on the distance.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--TRAP IF POSSIBLE</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; i<span class="sy0">=</span><span class="nu0">1</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">while</span> <span class="br0">&#40;</span>i<span class="sy0">&lt;=</span>nNumMinions<span class="br0">&#41;</span> <span class="kw1">do</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span> <span class="kw2">and</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>IsValid<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>IsAlive<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; vecTrappingPosition <span class="sy0">=</span> positionOffset<span class="br0">&#40;</span>vecTargetPos<span class="sy0">,</span> angleToWell <span class="sy0">+</span><span class="nu0">2</span><span class="sy0">*</span>pi<span class="sy0">*</span>i<span class="sy0">/</span>nNumMinions<span class="sy0">,</span> distance<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; nDistFromTargetSq<span class="sy0">=</span>Vector3<span class="sy0">.</span>Distance2DSq<span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> vecTargetPos<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- Don't try to trap if you or they are unitwalking, or trapping already.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>target<span class="sy0">:</span>GetUnitwalking<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">or</span> HoN<span class="sy0">:</span>GetGameTime<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">-</span>object<span class="sy0">.</span>nTimeEnergizered<span class="sy0">&lt;</span><span class="nu0">5200</span> <span class="kw2">or</span> <span class="br0">&#40;</span>HoN<span class="sy0">:</span>GetGameTime<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">-</span>object<span class="sy0">.</span>nTimeEnergizered<span class="sy0">&gt;</span><span class="nu0">6000</span> <span class="kw2">and</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>GetUnitwalking<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="kw2">or</span> target<span class="sy0">:</span>GetHealthPercent<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">*</span><span class="nu0">100</span><span class="sy0">&lt;</span><span class="nu0">5</span> <span class="kw2">or</span> <span class="br0">&#40;</span>Vector3<span class="sy0">.</span>Distance2DSq<span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> vecTrappingPosition<span class="br0">&#41;</span><span class="sy0">&lt;</span>object<span class="sy0">.</span>trapEffectiveness<span class="br0">&#91;</span>core<span class="sy0">.</span>nDifficulty<span class="br0">&#93;</span><span class="br0">&#41;</span> <span class="kw2">and</span> nDistFromTargetSq<span class="sy0">&lt;</span>object<span class="sy0">.</span>trapEffectiveness<span class="br0">&#91;</span>core<span class="sy0">.</span>nDifficulty<span class="br0">&#93;</span><span class="br0">&#41;</span> <span class="kw1">then</span><span class="co1">--35 was tmp</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">---------------------------------</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- &nbsp; &nbsp; &nbsp; a.Attack Target &nbsp; &nbsp; &nbsp; --</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">---------------------------------</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>nDistFromTargetSq<span class="sy0">&gt;</span><span class="nu0">128</span><span class="sy0">*</span><span class="nu0">128</span><span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--ATTACK, You are/can't trap them.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>core<span class="sy0">.</span>GetAttackSequenceProgress<span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="br0">&#41;</span><span class="sy0">~=</span><span class="st0">&quot;windup&quot;</span><span class="br0">&#41;</span> <span class="kw1">then</span><span class="co1">--this is needed, don't cancel attacks param doesn't work for some reason.</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; core<span class="sy0">.</span>OrderMoveToPos<span class="br0">&#40;</span>botBrain<span class="sy0">,</span>object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">,</span> vecTargetPos<span class="sy0">,</span><span class="kw4">false</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>bDebugLines<span class="br0">&#41;</span><span class="kw1">then</span> core<span class="sy0">.</span>DrawXPosition<span class="br0">&#40;</span>vecTargetPos<span class="sy0">,</span> <span class="st0">'yellow'</span><span class="br0">&#41;</span> <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">else</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; core<span class="sy0">.</span>OrderAttack<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">,</span> target<span class="sy0">,</span><span class="kw4">false</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>bDebugLines<span class="br0">&#41;</span><span class="kw1">then</span> core<span class="sy0">.</span>DrawXPosition<span class="br0">&#40;</span>vecTargetPos<span class="sy0">,</span> <span class="st0">'red'</span><span class="br0">&#41;</span> <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">else</span> <span class="co1">--TRAP, they are trappable.</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">---------------------------------</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- &nbsp; &nbsp; &nbsp;b.Surround Target &nbsp; &nbsp; &nbsp;--</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">---------------------------------</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; botBrain<span class="sy0">:</span>OrderPosition<span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">.</span>object <span class="kw2">or</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">,</span> <span class="st0">&quot;Move&quot;</span><span class="sy0">,</span>vecTrappingPosition<span class="sy0">,</span><span class="kw4">nil</span><span class="sy0">,</span> <span class="kw4">nil</span><span class="sy0">,</span><span class="kw4">false</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>bDebugLines<span class="br0">&#41;</span><span class="kw1">then</span> core<span class="sy0">.</span>DrawDebugArrow<span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> vecTrappingPosition<span class="sy0">,</span> <span class="st0">'green'</span><span class="br0">&#41;</span> <span class="kw1">end</span><span class="co1">--draw where hero is going</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; i<span class="sy0">=</span>i<span class="sy0">+</span><span class="nu0">1</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">else</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw3">table.remove</span><span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="sy0">,</span>i<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; nNumMinions<span class="sy0">=</span><span class="kw3">table.getn</span><span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">else</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">---------------------------------</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- &nbsp; &nbsp; &nbsp; c.Attack &nbsp;Target &nbsp; &nbsp; &nbsp;--</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">---------------------------------</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--we don't have enough minions to trap, just attack.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; object<span class="sy0">.</span>nMinionSkip<span class="sy0">=</span><span class="nu0">5</span><span class="co1">--250ms till next cycle</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">for</span> i<span class="sy0">=</span><span class="nu0">1</span><span class="sy0">,</span>nNumMinions <span class="kw1">do</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span> <span class="kw2">and</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>IsValid<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>IsAlive<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="kw1">then</span> core<span class="sy0">.</span>OrderAttack<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">,</span> target<span class="br0">&#41;</span> <span class="kw1">end</span> <span class="co1">--kill em'</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">---------------------------------</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;4.Pushing &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;--</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">---------------------------------</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">else</span> <span class="co1">--no target, push forwards.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; object<span class="sy0">.</span>nMinionSkip<span class="sy0">=</span><span class="nu0">5</span><span class="co1">--250ms till next cycle</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; destPos<span class="sy0">=</span>core<span class="sy0">.</span>GetFurthestCreepWavePos<span class="br0">&#40;</span>core<span class="sy0">.</span>tMyLane<span class="sy0">,</span> <span class="kw4">true</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">for</span> i<span class="sy0">=</span><span class="nu0">1</span><span class="sy0">,</span>nNumMinions <span class="kw1">do</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="br0">&#41;</span><span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span> <span class="kw2">and</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>IsValid<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">:</span>IsAlive<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; core<span class="sy0">.</span>OrderAttackPosition<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> object<span class="sy0">.</span>tMinions<span class="br0">&#91;</span>i<span class="br0">&#93;</span><span class="sy0">,</span> destPos<span class="sy0">,</span><span class="kw4">false</span><span class="sy0">,</span><span class="kw4">false</span><span class="br0">&#41;</span><span class="co1">--attackmove</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">else</span> <span class="kw3">table.remove</span><span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="sy0">,</span>i<span class="br0">&#41;</span> i<span class="sy0">=</span>i<span class="sy0">-</span><span class="nu0">1</span> nNumMinions<span class="sy0">=</span><span class="kw3">table.getn</span><span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#41;</span> <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; nNumMinions<span class="sy0">=</span><span class="kw3">table.getn</span><span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1"><span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>onthinkOld <span class="sy0">=</span> object<span class="sy0">.</span>onthink</div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>onthink &nbsp;<span class="sy0">=</span> object<span class="sy0">.</span>onthinkOverride</div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1"><span class="co1">----------------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">-- &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;OnCombatEvent Override &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;--</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">----------------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">function</span> object<span class="sy0">:</span>oncombateventOverride<span class="br0">&#40;</span>EventData<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; self<span class="sy0">:</span>oncombateventOld<span class="br0">&#40;</span>EventData<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nAddBonus <span class="sy0">=</span> <span class="nu0">0</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> EventData<span class="sy0">.</span>Type <span class="sy0">==</span> <span class="st0">&quot;Ability&quot;</span> <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> EventData<span class="sy0">.</span>InflictorName <span class="sy0">==</span> <span class="st0">&quot;Ability_Bephelgor1&quot;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; nAddBonus <span class="sy0">=</span> nAddBonus <span class="sy0">+</span> object<span class="sy0">.</span>nRegurgitateUse</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">elseif</span> EventData<span class="sy0">.</span>InflictorName <span class="sy0">==</span> <span class="st0">&quot;Ability_Bephelgor4&quot;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; nAddBonus <span class="sy0">=</span> nAddBonus <span class="sy0">+</span> object<span class="sy0">.</span>nHellUse</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">elseif</span> EventData<span class="sy0">.</span>Type <span class="sy0">==</span> <span class="st0">&quot;Item&quot;</span> <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> core<span class="sy0">.</span>itemEnergizer <span class="sy0">~=</span> <span class="kw4">nil</span> <span class="kw2">and</span> EventData<span class="sy0">.</span>SourceUnit <span class="sy0">==</span> core<span class="sy0">.</span>unitSelf<span class="sy0">:</span>GetUniqueID<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> EventData<span class="sy0">.</span>InflictorName <span class="sy0">==</span> core<span class="sy0">.</span>itemEnergizer<span class="sy0">:</span>GetName<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; nAddBonus <span class="sy0">=</span> nAddBonus <span class="sy0">+</span> self<span class="sy0">.</span>nEnergizerUse</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> nAddBonus <span class="sy0">&gt;</span> <span class="nu0">0</span> <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; core<span class="sy0">.</span>DecayBonus<span class="br0">&#40;</span>self<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; core<span class="sy0">.</span>nHarassBonus <span class="sy0">=</span> core<span class="sy0">.</span>nHarassBonus <span class="sy0">+</span> nAddBonus</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1"><span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>oncombateventOld <span class="sy0">=</span> object<span class="sy0">.</span>oncombatevent</div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>oncombatevent <span class="sy0">=</span> object<span class="sy0">.</span>oncombateventOverride</div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1"><span class="co1">----------------------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">-- &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;CustomHarassUtility Override &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;--</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">----------------------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">local</span> <span class="kw1">function</span> CustomHarassUtilityOverride<span class="br0">&#40;</span>unit<span class="br0">&#41;</span> <span class="co1">--this is part of harass hero utility.</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; unitSelf<span class="sy0">=</span>core<span class="sy0">.</span>unitSelf</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> val <span class="sy0">=</span> <span class="nu0">0</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> heroTarget <span class="sy0">=</span> unit</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--can attack enemy</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>Vector3<span class="sy0">.</span>Distance2DSq<span class="br0">&#40;</span>unitSelf<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> heroTarget<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span><span class="sy0">&lt;</span><span class="nu0">128</span><span class="sy0">*</span><span class="nu0">128</span> <span class="kw2">and</span> unitSelf<span class="sy0">:</span>IsAttackReady<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="kw1">then</span> val<span class="sy0">=</span>val<span class="sy0">+</span>object<span class="sy0">.</span>nCanAttackUp <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--can use ultimate.</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> skills<span class="sy0">.</span>abilHellOnNewerth<span class="sy0">:</span>CanActivate<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw1">then</span> val <span class="sy0">=</span> val <span class="sy0">+</span> object<span class="sy0">.</span>nHellUp <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--can use barf</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> skills<span class="sy0">.</span>abilRegurgitate<span class="sy0">:</span>CanActivate<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw1">then</span> val <span class="sy0">=</span> val <span class="sy0">+</span> object<span class="sy0">.</span>nRegurgitateUp <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nNumMinions<span class="sy0">=#</span>object<span class="sy0">.</span>tMinions <span class="co1">--this is the same as table.getn(minions)</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--Minions are close to enemy</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>nNumMinions<span class="sy0">&gt;</span><span class="nu0">3</span> <span class="kw2">and</span> object<span class="sy0">.</span>nRealDistance <span class="kw2">and</span> object<span class="sy0">.</span>nRealDistance<span class="sy0">&lt;</span><span class="nu0">300</span><span class="br0">&#41;</span> <span class="kw1">then</span> val<span class="sy0">=</span>val<span class="sy0">+</span>object<span class="sy0">.</span>nMinionsCloseUp <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--Minions are trapping enemy</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>nNumMinions<span class="sy0">&gt;</span><span class="nu0">3</span> <span class="kw2">and</span> object<span class="sy0">.</span>bUnitStill <span class="kw2">and</span> object<span class="sy0">.</span>nRealDistance <span class="kw2">and</span> object<span class="sy0">.</span>nRealDistance<span class="sy0">&lt;</span><span class="nu0">80</span><span class="br0">&#41;</span><span class="kw1">then</span> val<span class="sy0">=</span>val<span class="sy0">+</span>object<span class="sy0">.</span>nTrappingUp <span class="kw1">end</span><span class="co1">--trapping! 10</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--can use energizer</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>core<span class="sy0">.</span>itemEnergizer <span class="kw2">and</span> object<span class="sy0">.</span>nMinionsClose<span class="sy0">&gt;</span><span class="nu0">3</span> <span class="kw2">and</span> core<span class="sy0">.</span>itemEnergizer<span class="sy0">:</span>CanActivate<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="kw1">then</span> val<span class="sy0">=</span>val<span class="sy0">+</span>object<span class="sy0">.</span>nEnergizerUp <span class="kw1">end</span> <span class="co1">--can activate energizer</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> val</div></li>
-<li class="li2"><div class="de2"><span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">behaviorLib<span class="sy0">.</span>CustomHarassUtility <span class="sy0">=</span> CustomHarassUtilityOverride</div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1"><span class="co1">----------------------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">-- &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; Last hitting Overrides &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; --</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">----------------------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">local</span> <span class="kw1">function</span> KaiGetCreepAttackTargetOverride<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> unitEnemyCreep<span class="sy0">,</span> unitAllyCreep<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> unitSelf <span class="sy0">=</span> core<span class="sy0">.</span>unitSelf</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--Get closest hero. We can make decisions based on this.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> uClosestHero <span class="sy0">=</span> <span class="kw4">nil</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nClosestHeroDistSq <span class="sy0">=</span> <span class="nu0">2000</span><span class="sy0">*</span><span class="nu0">2000</span> <span class="co1">-- more than 2000 and we are not concerned. 2000*2000</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">for</span> id<span class="sy0">,</span> unitHero <span class="kw2">in</span> <span class="kw3">pairs</span><span class="br0">&#40;</span>HoN<span class="sy0">.</span>GetHeroes<span class="br0">&#40;</span>core<span class="sy0">.</span>enemyTeam<span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="kw1">do</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> unitHero <span class="sy0">~=</span> <span class="kw4">nil</span> <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> core<span class="sy0">.</span>CanSeeUnit<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> unitHero<span class="br0">&#41;</span> <span class="kw2">and</span> unitHero<span class="sy0">:</span>GetTeam<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">~=</span>team <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nDistanceSq <span class="sy0">=</span> Vector3<span class="sy0">.</span>Distance2DSq<span class="br0">&#40;</span>unitHero<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> core<span class="sy0">.</span>unitSelf<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> nDistanceSq <span class="sy0">&lt;</span> nClosestHeroDistSq <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; nClosestHeroDistSq <span class="sy0">=</span> nDistanceSq</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; uClosestHero <span class="sy0">=</span> unitHero</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--addition is how much sooner to move towards creep. '30' is 30hp more than you can 1 hit kill it for.</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; addition<span class="sy0">=</span><span class="nu0">30</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>uClosestHero <span class="kw2">and</span> uClosestHero<span class="sy0">:</span>GetAttackType<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="sy0">~=</span> <span class="st0">&quot;melee&quot;</span><span class="br0">&#41;</span> <span class="kw1">then</span> <span class="co1">--be more passive if opponent is ranged.</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; addition<span class="sy0">=</span><span class="nu0">20</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span> &nbsp; &nbsp; </div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>uClosestHero <span class="kw2">and</span> unitSelf<span class="sy0">:</span>GetHealthPercent<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">*</span><span class="nu0">100</span><span class="sy0">&lt;</span><span class="nu0">60</span> <span class="kw2">and</span> Vector3<span class="sy0">.</span>Distance2D<span class="br0">&#40;</span>unitSelf<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> uClosestHero<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span><span class="sy0">&lt;</span>uClosestHero<span class="sy0">:</span>GetAttackRange<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">+</span><span class="nu0">150</span><span class="br0">&#41;</span> <span class="kw1">then</span> <span class="kw1">return</span> <span class="kw4">nil</span> <span class="kw1">end</span> <span class="co1">--too dangerous. Stay back.</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nDamageAverage <span class="sy0">=</span> unitSelf<span class="sy0">:</span>GetFinalAttackDamageMin<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">+</span>addition <span class="co1">--make the hero go to the unit when it is 40 hp away, but too risky if enemy is ranged</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; core<span class="sy0">.</span>FindItems<span class="br0">&#40;</span>botBrain<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> core<span class="sy0">.</span>itemHatchet <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; nDamageAverage <span class="sy0">=</span> nDamageAverage <span class="sy0">*</span> core<span class="sy0">.</span>itemHatchet<span class="sy0">.</span>creepDamageMul</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> core<span class="sy0">.</span>nDifficulty <span class="sy0">==</span> core<span class="sy0">.</span>nEASY_DIFFICULTY <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; nDamageAverage <span class="sy0">=</span> nDamageAverage <span class="sy0">+</span> <span class="nu0">10</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> unitEnemyCreep <span class="kw2">and</span> core<span class="sy0">.</span>CanSeeUnit<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> unitEnemyCreep<span class="br0">&#41;</span> <span class="kw1">then</span> <span class="co1">--return enemy creep</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nTargetHealth <span class="sy0">=</span> unitEnemyCreep<span class="sy0">:</span>GetHealth<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> nDamageAverage <span class="sy0">&gt;=</span> nTargetHealth <span class="kw2">or</span> core<span class="sy0">.</span>teamBotBrain<span class="sy0">.</span>nPushState<span class="sy0">==</span><span class="nu0">2</span> <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> unitEnemyCreep</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> unitAllyCreep <span class="kw2">and</span> core<span class="sy0">.</span>nDifficulty <span class="sy0">~=</span> core<span class="sy0">.</span>nEASY_DIFFICULTY <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nTargetHealth <span class="sy0">=</span> unitAllyCreep<span class="sy0">:</span>GetHealth<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> nDamageAverage <span class="sy0">&gt;=</span> nTargetHealth <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> core<span class="sy0">.</span>teamBotBrain<span class="sy0">.</span>nPushState<span class="sy0">~=</span><span class="nu0">2</span> <span class="kw2">and</span> <span class="br0">&#40;</span>core<span class="sy0">.</span>nDifficulty <span class="sy0">~=</span> core<span class="sy0">.</span>nEASY_DIFFICULTY <span class="kw2">or</span> <span class="br0">&#40;</span>core<span class="sy0">.</span>bIsTutorial <span class="kw2">and</span> core<span class="sy0">.</span>bTutorialBehaviorReset <span class="sy0">==</span> <span class="kw4">true</span> <span class="kw2">and</span> core<span class="sy0">.</span>myTeam <span class="sy0">==</span> HoN<span class="sy0">.</span>GetHellbourneTeam<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> unitAllyCreep</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> <span class="kw4">nil</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>GetCreepAttackTargetOld <span class="sy0">=</span> behaviorLib<span class="sy0">.</span>GetCreepAttackTarget</div></li>
-<li class="li2"><div class="de2">behaviorLib<span class="sy0">.</span>GetCreepAttackTarget <span class="sy0">=</span> KaiGetCreepAttackTargetOverride</div></li>
-<li class="li1"><div class="de1">&nbsp;</div></li>
-<li class="li2"><div class="de2"><span class="kw1">local</span> <span class="kw1">function</span> KaiAttackCreepsExecuteOverride<span class="br0">&#40;</span>botBrain<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> unitSelf <span class="sy0">=</span> core<span class="sy0">.</span>unitSelf</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> currentTarget <span class="sy0">=</span> core<span class="sy0">.</span>unitCreepTarget</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> currentTarget <span class="kw2">and</span> core<span class="sy0">.</span>CanSeeUnit<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> currentTarget<span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> vecTargetPos <span class="sy0">=</span> currentTarget<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nDistSq <span class="sy0">=</span> Vector3<span class="sy0">.</span>Distance2DSq<span class="br0">&#40;</span>unitSelf<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> vecTargetPos<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nAttackRangeSq <span class="sy0">=</span> core<span class="sy0">.</span>GetAbsoluteAttackRangeToUnit<span class="br0">&#40;</span>unitSelf<span class="sy0">,</span> currentTarget<span class="sy0">,</span> <span class="kw4">true</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nDamageAverage <span class="sy0">=</span> unitSelf<span class="sy0">:</span>GetFinalAttackDamageMin<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> core<span class="sy0">.</span>itemHatchet <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; nDamageAverage <span class="sy0">=</span> nDamageAverage <span class="sy0">*</span> core<span class="sy0">.</span>itemHatchet<span class="sy0">.</span>creepDamageMul</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> currentTarget <span class="sy0">~=</span> <span class="kw4">nil</span> <span class="kw1">then</span>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> nDistSq <span class="sy0">&lt;</span> nAttackRangeSq <span class="kw2">and</span> unitSelf<span class="sy0">:</span>IsAttackReady<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> <span class="br0">&#40;</span>nDamageAverage<span class="sy0">*</span><span class="br0">&#40;</span><span class="nu0">1</span><span class="sy0">-</span>currentTarget<span class="sy0">:</span>GetPhysicalResistance<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span><span class="sy0">&gt;=</span>currentTarget<span class="sy0">:</span>GetHealth<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">or</span> core<span class="sy0">.</span>teamBotBrain<span class="sy0">.</span>nPushState<span class="sy0">==</span><span class="nu0">2</span><span class="br0">&#41;</span> <span class="kw1">then</span> <span class="co1">--only kill if you can get gold</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> core<span class="sy0">.</span>OrderAttackClamp<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> unitSelf<span class="sy0">,</span> currentTarget<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">elseif</span> nDistSq <span class="sy0">&gt;</span> nAttackRangeSq <span class="kw2">and</span> vecTargetPos <span class="kw2">and</span> unitSelf<span class="sy0">:</span>IsAttackReady<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> <span class="br0">&#40;</span>nDamageAverage<span class="sy0">*</span><span class="br0">&#40;</span><span class="nu0">1</span><span class="sy0">-</span>currentTarget<span class="sy0">:</span>GetPhysicalResistance<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span><span class="sy0">&gt;=</span>currentTarget<span class="sy0">:</span>GetHealth<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">-</span><span class="nu0">100</span> <span class="kw2">or</span> core<span class="sy0">.</span>teamBotBrain<span class="sy0">.</span>nPushState<span class="sy0">==</span><span class="nu0">2</span><span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> core<span class="sy0">.</span>OrderMoveToPosClamp<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> unitSelf<span class="sy0">,</span> vecTargetPos<span class="br0">&#41;</span> <span class="co1">--moves hero to target</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">else</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> core<span class="sy0">.</span>OrderHoldClamp<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> unitSelf<span class="sy0">,</span> <span class="kw4">false</span><span class="br0">&#41;</span> <span class="co1">--this is where the magic happens. Wait for the kill.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> <span class="kw4">false</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>AttackCreepsExecuteOld <span class="sy0">=</span> behaviorLib<span class="sy0">.</span>AttackCreepsBehavior<span class="br0">&#91;</span><span class="st0">&quot;Execute&quot;</span><span class="br0">&#93;</span></div></li>
-<li class="li2"><div class="de2">behaviorLib<span class="sy0">.</span>AttackCreepsBehavior<span class="br0">&#91;</span><span class="st0">&quot;Execute&quot;</span><span class="br0">&#93;</span> <span class="sy0">=</span> KaiAttackCreepsExecuteOverride</div></li>
-<li class="li1"><div class="de1">&nbsp;</div></li>
-<li class="li2"><div class="de2"><span class="co1">----------------------------------------------------</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">-- &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; Heal At Well Override &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;--</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">----------------------------------------------------</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">--return to well more often. --2000 gold adds 8 to return utility, 0% mana also adds 8.</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">--When returning to well, use skills and items.</span></div></li>
-<li class="li1"><div class="de1"><span class="kw1">local</span> <span class="kw1">function</span> HealAtWellUtilityOverride<span class="br0">&#40;</span>botBrain<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>Vector3<span class="sy0">.</span>Distance2DSq<span class="br0">&#40;</span>core<span class="sy0">.</span>unitSelf<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> core<span class="sy0">.</span>allyWell <span class="kw2">and</span> core<span class="sy0">.</span>allyWell<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">or</span> behaviorLib<span class="sy0">.</span>PositionSelfBackUp<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span><span class="sy0">&lt;</span><span class="nu0">400</span><span class="sy0">*</span><span class="nu0">400</span> <span class="kw2">and</span> core<span class="sy0">.</span>unitSelf<span class="sy0">:</span>GetManaPercent<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">*</span><span class="nu0">100</span><span class="sy0">&lt;</span><span class="nu0">95</span><span class="br0">&#41;</span> <span class="kw1">then</span> <span class="kw1">return</span> <span class="nu0">80</span> <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> object<span class="sy0">.</span>HealAtWellUtilityOld<span class="br0">&#40;</span>botBrain<span class="br0">&#41;</span><span class="sy0">*</span><span class="nu0">1.75</span><span class="sy0">+</span><span class="br0">&#40;</span>botBrain<span class="sy0">:</span>GetGold<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">*</span><span class="nu0">8</span><span class="sy0">/</span><span class="nu0">2000</span><span class="br0">&#41;</span><span class="sy0">+</span> <span class="nu0">8</span><span class="sy0">-</span><span class="br0">&#40;</span>core<span class="sy0">.</span>unitSelf<span class="sy0">:</span>GetManaPercent<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">*</span><span class="nu0">8</span><span class="br0">&#41;</span> <span class="co1">--couragously flee back to base.</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1"><span class="kw1">local</span> <span class="kw1">function</span> HealAtWellExecuteOverride<span class="br0">&#40;</span>botBrain<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> vecWellPos <span class="sy0">=</span> core<span class="sy0">.</span>allyWell <span class="kw2">and</span> core<span class="sy0">.</span>allyWell<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">or</span> behaviorLib<span class="sy0">.</span>PositionSelfBackUp<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>Vector3<span class="sy0">.</span>Distance2DSq<span class="br0">&#40;</span>core<span class="sy0">.</span>unitSelf<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> vecWellPos<span class="br0">&#41;</span><span class="sy0">&gt;</span><span class="nu0">600</span><span class="sy0">*</span><span class="nu0">600</span><span class="br0">&#41;</span><span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>core<span class="sy0">.</span>itemEnergizer <span class="kw2">and</span> core<span class="sy0">.</span>itemEnergizer<span class="sy0">:</span>CanActivate<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> <span class="kw2">not</span> unitSelf<span class="sy0">:</span>HasState<span class="br0">&#40;</span><span class="st0">&quot;State_Energizer_Buff&quot;</span><span class="br0">&#41;</span><span class="br0">&#41;</span><span class="kw1">then</span> <span class="co1">--when heading to base, use energizer</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; botBrain<span class="sy0">:</span>OrderItem<span class="br0">&#40;</span>core<span class="sy0">.</span>itemEnergizer<span class="sy0">.</span>object <span class="kw2">or</span> core<span class="sy0">.</span>itemEnergizer<span class="sy0">,</span> <span class="kw4">false</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; object<span class="sy0">.</span>nTimeEnergizered<span class="sy0">=</span>HoN<span class="sy0">:</span>GetGameTime<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>core<span class="sy0">.</span>itemGhostMarchers <span class="kw2">and</span> core<span class="sy0">.</span>itemGhostMarchers<span class="sy0">:</span>CanActivate<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span><span class="kw1">then</span> <span class="co1">--when heading to base, use boots</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; botBrain<span class="sy0">:</span>OrderItem<span class="br0">&#40;</span>core<span class="sy0">.</span>itemGhostMarchers<span class="sy0">.</span>object <span class="kw2">or</span> core<span class="sy0">.</span>itemGhostMarchers<span class="sy0">,</span> <span class="kw4">false</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>skills<span class="sy0">.</span>abilRegurgitate<span class="sy0">:</span>CanActivate<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="kw1">then</span> <span class="co1">--BARF</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; core<span class="sy0">.</span>OrderAbility<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> skills<span class="sy0">.</span>abilRegurgitate<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> object<span class="sy0">.</span>HealAtWellExecuteOld<span class="br0">&#40;</span>botBrain<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">object<span class="sy0">.</span>HealAtWellUtilityOld <span class="sy0">=</span> behaviorLib<span class="sy0">.</span>HealAtWellBehavior<span class="br0">&#91;</span><span class="st0">&quot;Utility&quot;</span><span class="br0">&#93;</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>HealAtWellExecuteOld <span class="sy0">=</span> behaviorLib<span class="sy0">.</span>HealAtWellBehavior<span class="br0">&#91;</span><span class="st0">&quot;Execute&quot;</span><span class="br0">&#93;</span></div></li>
-<li class="li1"><div class="de1">behaviorLib<span class="sy0">.</span>HealAtWellBehavior<span class="br0">&#91;</span><span class="st0">&quot;Utility&quot;</span><span class="br0">&#93;</span> <span class="sy0">=</span> HealAtWellUtilityOverride</div></li>
-<li class="li2"><div class="de2">behaviorLib<span class="sy0">.</span>HealAtWellBehavior<span class="br0">&#91;</span><span class="st0">&quot;Execute&quot;</span><span class="br0">&#93;</span> <span class="sy0">=</span> HealAtWellExecuteOverride</div></li>
-<li class="li1"><div class="de1">&nbsp;</div></li>
-<li class="li2"><div class="de2"><span class="co1">----------------------------------------</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">-- &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;Harass Behaviour &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;--</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">----------------------------------------</span></div></li>
-<li class="li1"><div class="de1"><span class="kw1">local</span> <span class="kw1">function</span> HarassHeroExecuteOverride<span class="br0">&#40;</span>botBrain<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> unitSelf <span class="sy0">=</span> core<span class="sy0">.</span>unitSelf</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nHarassUtility <span class="sy0">=</span> behaviorLib<span class="sy0">.</span>lastHarassUtil</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> unitTarget <span class="sy0">=</span> behaviorLib<span class="sy0">.</span>heroTarget</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nTargetDistanceSq <span class="sy0">=</span> Vector3<span class="sy0">.</span>Distance2DSq<span class="br0">&#40;</span>unitSelf<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> unitTarget<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nNumMinions<span class="sy0">=</span><span class="kw3">table.getn</span><span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> vecWellPos <span class="sy0">=</span> core<span class="sy0">.</span>allyWell <span class="kw2">and</span> core<span class="sy0">.</span>allyWell<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">or</span> behaviorLib<span class="sy0">.</span>PositionSelfBackUp<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> distToWellSq<span class="sy0">=</span>Vector3<span class="sy0">.</span>Distance2DSq<span class="br0">&#40;</span>unitSelf<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> vecWellPos<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> bActionTaken <span class="sy0">=</span> <span class="kw4">false</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--Spawn Minions</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> skills<span class="sy0">.</span>abilSpawnMinions<span class="sy0">:</span>CanActivate<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> distToWellSq<span class="sy0">&gt;</span><span class="nu0">500</span><span class="sy0">*</span><span class="nu0">500</span> <span class="kw2">and</span> unitSelf<span class="sy0">:</span>GetLevel<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">&gt;=</span><span class="nu0">5</span> <span class="kw2">and</span> skills<span class="sy0">.</span>abilSpawnMinions<span class="sy0">:</span>GetCharges<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">==</span>skills<span class="sy0">.</span>abilSpawnMinions<span class="sy0">:</span>GetLevel<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">*</span><span class="nu0">3</span> <span class="kw2">and</span> nNumMinions<span class="sy0">&lt;</span><span class="nu0">3</span> <span class="kw2">and</span> <span class="co1">--can activate</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="br0">&#40;</span>core<span class="sy0">.</span>teamBotBrain<span class="sy0">.</span>nPushState<span class="sy0">==</span><span class="nu0">2</span> <span class="kw2">or</span> behaviorLib<span class="sy0">.</span>heroTarget<span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; bActionTaken<span class="sy0">=</span>core<span class="sy0">.</span>OrderAbility<span class="br0">&#40;</span>botBrain<span class="sy0">,</span>skills<span class="sy0">.</span>abilSpawnMinions<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--Regurgitate</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="kw2">not</span> bActionTaken <span class="kw2">and</span> nHarassUtility<span class="sy0">&gt;</span>object<span class="sy0">.</span>nRegurgitateThreshold <span class="kw2">and</span> nTargetDistanceSq<span class="sy0">&lt;</span><span class="nu0">300</span><span class="sy0">*</span><span class="nu0">300</span> <span class="kw2">and</span> skills<span class="sy0">.</span>abilRegurgitate<span class="sy0">:</span>CanActivate<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; object<span class="sy0">.</span>nTimeBarfed<span class="sy0">=</span>HoN<span class="sy0">:</span>GetGameTime<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; bActionTaken <span class="sy0">=</span> core<span class="sy0">.</span>OrderAbility<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> skills<span class="sy0">.</span>abilRegurgitate<span class="sy0">,</span><span class="kw4">false</span><span class="sy0">,</span><span class="kw4">false</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- Ultimate</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> ultThreshholdValue<span class="sy0">=</span><span class="kw3">floor</span><span class="br0">&#40;</span>skills<span class="sy0">.</span>abilHellOnNewerth<span class="sy0">:</span>GetCharges<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">/</span><span class="nu0">40</span><span class="br0">&#41;</span><span class="sy0">+</span><span class="nu0">1</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="kw2">not</span> bActionTaken <span class="kw2">and</span> skills<span class="sy0">.</span>abilHellOnNewerth<span class="sy0">:</span>CanActivate<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> nHarassUtility<span class="sy0">&gt;</span>object<span class="sy0">.</span>tHellOnNewerthThreshold<span class="br0">&#91;</span>ultThreshholdValue<span class="br0">&#93;</span> <span class="kw2">and</span> nTargetDistanceSq<span class="sy0">&lt;</span><span class="nu0">550</span><span class="sy0">*</span><span class="nu0">550</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; bActionTaken <span class="sy0">=</span> core<span class="sy0">.</span>OrderAbility<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> skills<span class="sy0">.</span>abilHellOnNewerth<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- Energizer</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="kw2">not</span> bActionTaken <span class="kw2">and</span> core<span class="sy0">.</span>itemEnergizer <span class="kw2">and</span> core<span class="sy0">.</span>itemEnergizer<span class="sy0">:</span>CanActivate<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> nHarassUtility<span class="sy0">&gt;</span>object<span class="sy0">.</span>nEnergizerThreshold <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; botBrain<span class="sy0">:</span>OrderItem<span class="br0">&#40;</span>core<span class="sy0">.</span>itemEnergizer<span class="sy0">.</span>object <span class="kw2">or</span> core<span class="sy0">.</span>itemEnergizer<span class="sy0">,</span> <span class="kw4">false</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; object<span class="sy0">.</span>nTimeEnergizered<span class="sy0">=</span>HoN<span class="sy0">:</span>GetGameTime<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--default</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="kw2">not</span> bActionTaken <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> object<span class="sy0">.</span>harassHeroExecuteOld<span class="br0">&#40;</span>botBrain<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> bActionTaken</div></li>
-<li class="li1"><div class="de1"><span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>harassHeroExecuteOld <span class="sy0">=</span> behaviorLib<span class="sy0">.</span>HarassHeroBehavior<span class="br0">&#91;</span><span class="st0">&quot;Execute&quot;</span><span class="br0">&#93;</span></div></li>
-<li class="li1"><div class="de1">behaviorLib<span class="sy0">.</span>HarassHeroBehavior<span class="br0">&#91;</span><span class="st0">&quot;Execute&quot;</span><span class="br0">&#93;</span> <span class="sy0">=</span> HarassHeroExecuteOverride</div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1"><span class="co1">----------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">-- &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;Retreat Behaviour &nbsp; &nbsp; &nbsp; &nbsp; --</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">----------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">local</span> <span class="kw1">function</span> retreatFromThreatExecuteOverride<span class="br0">&#40;</span>botBrain<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--note that these actions do not requite a frame to complete, therefore bActionTaken is not necessary.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--use energizer when retreating</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> itemEnergizer <span class="kw2">and</span> itemEnergizer<span class="sy0">:</span>CanActivate<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> behaviorLib<span class="sy0">.</span>lastRetreatUtil<span class="sy0">&gt;=</span>object<span class="sy0">.</span>nEnergizerThreshold <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; botBrain<span class="sy0">:</span>OrderItem<span class="br0">&#40;</span>core<span class="sy0">.</span>itemEnergizer<span class="sy0">.</span>object <span class="kw2">or</span> core<span class="sy0">.</span>itemEnergizer<span class="sy0">,</span> <span class="kw4">false</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; object<span class="sy0">.</span>nTimeEnergizered<span class="sy0">=</span>HoN<span class="sy0">:</span>GetGameTime<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--use regurgitate when retreating</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>skills<span class="sy0">.</span>abilRegurgitate<span class="sy0">:</span>CanActivate<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> behaviorLib<span class="sy0">.</span>lastRetreatUtil<span class="sy0">&gt;=</span>object<span class="sy0">.</span>nRegurgitateThreshold<span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; core<span class="sy0">.</span>OrderAbility<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> skills<span class="sy0">.</span>abilRegurgitate<span class="sy0">,</span><span class="kw4">false</span><span class="sy0">,</span><span class="kw4">true</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--use ghost marchers when retreating</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>core<span class="sy0">.</span>itemGhostMarchers <span class="kw2">and</span> core<span class="sy0">.</span>itemGhostMarchers<span class="sy0">:</span>CanActivate<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span><span class="kw1">then</span> <span class="co1">--when heading to base, use boots</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; botBrain<span class="sy0">:</span>OrderItem<span class="br0">&#40;</span>core<span class="sy0">.</span>itemGhostMarchers<span class="sy0">.</span>object <span class="kw2">or</span> core<span class="sy0">.</span>itemGhostMarchers<span class="sy0">,</span> <span class="kw4">false</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> object<span class="sy0">.</span>RetreatFromThreatExecuteOld<span class="br0">&#40;</span>botBrain<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1"><span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">object<span class="sy0">.</span>RetreatFromThreatExecuteOld <span class="sy0">=</span> behaviorLib<span class="sy0">.</span>RetreatFromThreatBehavior<span class="br0">&#91;</span><span class="st0">&quot;Execute&quot;</span><span class="br0">&#93;</span></div></li>
-<li class="li1"><div class="de1">behaviorLib<span class="sy0">.</span>RetreatFromThreatBehavior<span class="br0">&#91;</span><span class="st0">&quot;Execute&quot;</span><span class="br0">&#93;</span> <span class="sy0">=</span> retreatFromThreatExecuteOverride</div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1"><span class="co1">----------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">-- &nbsp; &nbsp; &nbsp;Advanced Think Behaviour &nbsp; &nbsp; &nbsp;--</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">----------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">local</span> <span class="kw1">function</span> advancedThinkUtility<span class="br0">&#40;</span>botBrain<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> <span class="nu0">95</span><span class="sy0">;</span> <span class="co1">--always ridiculously important. Though, this rarly returns true.</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp;</div></li>
-<li class="li2"><div class="de2"><span class="kw1">local</span> <span class="kw1">function</span> advancedThinkExecute<span class="br0">&#40;</span>botBrain<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> unitSelf <span class="sy0">=</span> core<span class="sy0">.</span>unitSelf</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> vecSelfPos <span class="sy0">=</span> unitSelf<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nNumMinions<span class="sy0">=</span><span class="kw3">table.getn</span><span class="br0">&#40;</span>object<span class="sy0">.</span>tMinions<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> vecWellPos <span class="sy0">=</span> core<span class="sy0">.</span>allyWell <span class="kw2">and</span> core<span class="sy0">.</span>allyWell<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">or</span> behaviorLib<span class="sy0">.</span>PositionSelfBackUp<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> distToWellSq<span class="sy0">=</span>Vector3<span class="sy0">.</span>Distance2DSq<span class="br0">&#40;</span>vecSelfPos<span class="sy0">,</span> vecWellPos<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> bActionTaken<span class="sy0">=</span><span class="kw4">false</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--make and eat corpses at base</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span> distToWellSq<span class="sy0">&lt;</span><span class="nu0">500</span><span class="sy0">*</span><span class="nu0">500</span> <span class="kw2">and</span> <span class="kw2">not</span> unitSelf<span class="sy0">:</span>IsChanneling<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; unitSelf<span class="sy0">:</span>GetLevel<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">&gt;</span><span class="nu0">1</span> <span class="kw2">and</span> <span class="br0">&#40;</span>skills<span class="sy0">.</span>abilSpawnMinions<span class="sy0">:</span>GetCharges<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">&lt;</span>skills<span class="sy0">.</span>abilSpawnMinions<span class="sy0">:</span>GetLevel<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">*</span><span class="nu0">3</span> <span class="kw2">or</span> object<span class="sy0">.</span>nTimeBarfed<span class="sy0">+</span><span class="nu0">4800</span><span class="sy0">&gt;</span>HoN<span class="sy0">:</span>GetGameTime<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>skills<span class="sy0">.</span>abilRegurgitate<span class="sy0">:</span>CanActivate<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> <span class="kw2">not</span> unitSelf<span class="sy0">:</span>IsChanneling<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> distToWellSq<span class="sy0">&lt;</span><span class="nu0">100</span><span class="sy0">*</span><span class="nu0">100</span><span class="br0">&#41;</span> <span class="kw1">then</span> <span class="co1">--turn and spew</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; bActionTaken<span class="sy0">=</span>core<span class="sy0">.</span>OrderMoveToPos<span class="br0">&#40;</span>botBrain<span class="sy0">,</span>unitSelf<span class="sy0">,</span> positionOffset<span class="br0">&#40;</span>vecSelfPos<span class="sy0">,</span> <span class="nu0">0</span><span class="sy0">,</span> <span class="nu0">10</span><span class="br0">&#41;</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; object<span class="sy0">.</span>nTimeBarfed<span class="sy0">=</span>HoN<span class="sy0">:</span>GetGameTime<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; bActionTaken<span class="sy0">=</span>core<span class="sy0">.</span>OrderAbility<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> skills<span class="sy0">.</span>abilRegurgitate<span class="sy0">,</span><span class="kw4">false</span><span class="sy0">,</span><span class="kw4">true</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">elseif</span> <span class="br0">&#40;</span>object<span class="sy0">.</span>nTimeBarfed<span class="sy0">+</span><span class="nu0">3300</span><span class="sy0">&gt;</span>HoN<span class="sy0">:</span>GetGameTime<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="kw1">then</span> <span class="co1">--wait while barfing.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; bActionTaken<span class="sy0">=</span>core<span class="sy0">.</span>OrderHoldClamp<span class="br0">&#40;</span>botBrain<span class="sy0">,</span> unitSelf<span class="sy0">,</span> <span class="kw4">false</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">elseif</span> <span class="br0">&#40;</span>object<span class="sy0">.</span>nTimeBarfed<span class="sy0">+</span><span class="nu0">4200</span><span class="sy0">&gt;</span>HoN<span class="sy0">:</span>GetGameTime<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="kw1">then</span> <span class="co1">--collect corpses</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; bActionTaken<span class="sy0">=</span>core<span class="sy0">.</span>OrderMoveToPos<span class="br0">&#40;</span>botBrain<span class="sy0">,</span>unitSelf<span class="sy0">,</span> positionOffset<span class="br0">&#40;</span>vecSelfPos<span class="sy0">,</span> <span class="nu0">0</span><span class="sy0">,</span> <span class="nu0">300</span><span class="br0">&#41;</span><span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">else</span> <span class="co1">--move back into the center of the well</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; bActionTaken<span class="sy0">=</span>core<span class="sy0">.</span>OrderMoveToPos<span class="br0">&#40;</span>botBrain<span class="sy0">,</span>unitSelf<span class="sy0">,</span> vecWellPos<span class="br0">&#41;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--remove invalid hatchet item.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> core<span class="sy0">.</span>itemHatchet <span class="sy0">~=</span> <span class="kw4">nil</span> <span class="kw2">and</span> <span class="kw2">not</span> core<span class="sy0">.</span>itemHatchet<span class="sy0">:</span>IsValid<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; core<span class="sy0">.</span>itemHatchet <span class="sy0">=</span> <span class="kw4">nil</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--hatchet</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="kw2">not</span> bActionTaken <span class="kw2">and</span> core<span class="sy0">.</span>unitCreepTarget <span class="kw2">and</span> core<span class="sy0">.</span>itemHatchet <span class="kw2">and</span> core<span class="sy0">.</span>itemHatchet<span class="sy0">:</span>CanActivate<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> <span class="co1">--can activate</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; Vector3<span class="sy0">.</span>Distance2DSq<span class="br0">&#40;</span>unitSelf<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> core<span class="sy0">.</span>unitCreepTarget<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="sy0">&lt;=</span> <span class="nu0">600</span><span class="sy0">*</span><span class="nu0">600</span> <span class="kw2">and</span> <span class="co1">--in range of hatchet.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; unitSelf<span class="sy0">:</span>GetBaseDamage<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">*</span><span class="br0">&#40;</span><span class="nu0">1</span><span class="sy0">-</span>core<span class="sy0">.</span>unitCreepTarget<span class="sy0">:</span>GetPhysicalResistance<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span><span class="sy0">&gt;</span>core<span class="sy0">.</span>unitCreepTarget<span class="sy0">:</span>GetHealth<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> <span class="co1">--low enough hp, killable</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw3">string.find</span><span class="br0">&#40;</span>core<span class="sy0">.</span>unitCreepTarget<span class="sy0">:</span>GetTypeName<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> <span class="st0">&quot;Creep&quot;</span><span class="br0">&#41;</span> <span class="kw1">then</span><span class="co1">-- viable creep (this makes it ignore minions etc, some of which aren't hatchetable.)</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; botBrain<span class="sy0">:</span>OrderItemEntity<span class="br0">&#40;</span>core<span class="sy0">.</span>itemHatchet<span class="sy0">.</span>object <span class="kw2">or</span> core<span class="sy0">.</span>itemHatchet<span class="sy0">,</span> core<span class="sy0">.</span>unitCreepTarget<span class="sy0">.</span>object <span class="kw2">or</span> core<span class="sy0">.</span>unitCreepTarget<span class="sy0">,</span> <span class="kw4">false</span><span class="br0">&#41;</span><span class="co1">--use hatchet.</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> bActionTaken</div></li>
-<li class="li2"><div class="de2"><span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">behaviorLib<span class="sy0">.</span>advancedThink <span class="sy0">=</span> <span class="br0">&#123;</span><span class="br0">&#125;</span></div></li>
-<li class="li2"><div class="de2">behaviorLib<span class="sy0">.</span>advancedThink<span class="br0">&#91;</span><span class="st0">&quot;Utility&quot;</span><span class="br0">&#93;</span> <span class="sy0">=</span> advancedThinkUtility</div></li>
-<li class="li1"><div class="de1">behaviorLib<span class="sy0">.</span>advancedThink<span class="br0">&#91;</span><span class="st0">&quot;Execute&quot;</span><span class="br0">&#93;</span> <span class="sy0">=</span> advancedThinkExecute</div></li>
-<li class="li2"><div class="de2">behaviorLib<span class="sy0">.</span>advancedThink<span class="br0">&#91;</span><span class="st0">&quot;Name&quot;</span><span class="br0">&#93;</span> <span class="sy0">=</span> <span class="st0">&quot;advancedThink&quot;</span></div></li>
-<li class="li1"><div class="de1"><span class="kw3">tinsert</span><span class="br0">&#40;</span>behaviorLib<span class="sy0">.</span>tBehaviors<span class="sy0">,</span> behaviorLib<span class="sy0">.</span>advancedThink<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1"><span class="co1">----------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="co1">-- &nbsp; &nbsp; Collect Corpses Behaviour &nbsp; &nbsp; &nbsp;--</span></div></li>
-<li class="li1"><div class="de1"><span class="co1">----------------------------------------</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">local</span> corpsePosition<span class="sy0">=</span><span class="kw4">nil</span></div></li>
-<li class="li1"><div class="de1"><span class="kw1">local</span> <span class="kw1">function</span> getCorpsesUtility<span class="br0">&#40;</span>botBrain<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> unitSelf<span class="sy0">=</span>core<span class="sy0">.</span>unitSelf</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> vecSelfPos<span class="sy0">=</span>unitSelf<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; </div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">-- Find Corpses</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>unitSelf<span class="sy0">:</span>IsAlive<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> core<span class="sy0">.</span>localUnits<span class="sy0">~=</span><span class="kw4">nil</span><span class="br0">&#41;</span><span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> localCreeps<span class="sy0">=</span><span class="br0">&#123;</span><span class="br0">&#125;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--Get all creeps nearby and put them into a single table.</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">for</span> k<span class="sy0">,</span>v <span class="kw2">in</span> <span class="kw3">pairs</span><span class="br0">&#40;</span>core<span class="sy0">.</span>localUnits<span class="br0">&#91;</span><span class="st0">'EnemyCreeps'</span><span class="br0">&#93;</span><span class="br0">&#41;</span> <span class="kw1">do</span> localCreeps<span class="br0">&#91;</span>k<span class="br0">&#93;</span> <span class="sy0">=</span> v <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">for</span> k<span class="sy0">,</span>v <span class="kw2">in</span> <span class="kw3">pairs</span><span class="br0">&#40;</span>core<span class="sy0">.</span>localUnits<span class="br0">&#91;</span><span class="st0">'AllyCreeps'</span><span class="br0">&#93;</span><span class="br0">&#41;</span> <span class="kw1">do</span> localCreeps<span class="br0">&#91;</span>k<span class="br0">&#93;</span> <span class="sy0">=</span> v <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">for</span> k<span class="sy0">,</span>v <span class="kw2">in</span> <span class="kw3">pairs</span><span class="br0">&#40;</span>localCreeps<span class="br0">&#41;</span> <span class="kw1">do</span> object<span class="sy0">.</span>tOldLocalCreeps<span class="br0">&#91;</span>k<span class="br0">&#93;</span> <span class="sy0">=</span> <span class="kw4">nil</span> <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">for</span> key<span class="sy0">,</span> unit <span class="kw2">in</span> <span class="kw3">pairs</span><span class="br0">&#40;</span>object<span class="sy0">.</span>tOldLocalCreeps<span class="br0">&#41;</span> <span class="kw1">do</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; id<span class="sy0">=</span>key</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>unit <span class="kw2">and</span> unit<span class="sy0">:</span>IsValid<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> unit<span class="sy0">:</span>GetHealth<span class="br0">&#40;</span><span class="br0">&#41;</span> <span class="kw2">and</span> unit<span class="sy0">:</span>GetHealth<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">&lt;=</span><span class="nu0">0</span> <span class="kw2">and</span> <span class="kw3">string.find</span><span class="br0">&#40;</span>unit<span class="sy0">:</span>GetTypeName<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> <span class="st0">&quot;Creep&quot;</span><span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="kw1">then</span> <span class="co1">--these cause awful amounts of 'corpses' constantly.</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>unit<span class="sy0">:</span>GetAttackRange<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">~=</span><span class="nu0">690</span><span class="br0">&#41;</span> <span class="kw1">then</span><span class="co1">--catapults don't have corpses..</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw3">table.insert</span><span class="br0">&#40;</span>object<span class="sy0">.</span>tCorpseTable<span class="sy0">,</span><span class="br0">&#123;</span>position<span class="sy0">=</span>unit<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span><span class="kw3">time</span><span class="sy0">=</span>HoN<span class="sy0">:</span>GetGameTime<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#125;</span><span class="br0">&#41;</span><span class="co1">--create corpse record</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; object<span class="sy0">.</span>tOldLocalCreeps<span class="sy0">=</span>localCreeps</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">for</span> key<span class="sy0">,</span> event <span class="kw2">in</span> <span class="kw3">pairs</span><span class="br0">&#40;</span>object<span class="sy0">.</span>tCorpseTable<span class="br0">&#41;</span> <span class="kw1">do</span> <span class="co1">--remove old/obtained corpses</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>HoN<span class="sy0">.</span>GetGameTime<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">&gt;</span>event<span class="br0">&#91;</span><span class="st0">'time'</span><span class="br0">&#93;</span><span class="sy0">+</span><span class="nu0">14000</span> <span class="kw2">or</span> Vector3<span class="sy0">.</span>Distance2DSq<span class="br0">&#40;</span>vecSelfPos<span class="sy0">,</span>event<span class="br0">&#91;</span><span class="st0">'position'</span><span class="br0">&#93;</span><span class="br0">&#41;</span><span class="sy0">&lt;=</span><span class="nu0">105</span><span class="sy0">*</span><span class="nu0">105</span><span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw3">table.remove</span><span class="br0">&#40;</span>object<span class="sy0">.</span>tCorpseTable<span class="sy0">,</span>key<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp;</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--assess corpses</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> bDebugLines<span class="sy0">=</span><span class="kw4">false</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>unitSelf<span class="sy0">:</span>GetLevel<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">&lt;</span><span class="nu0">2</span> <span class="kw2">or</span> <span class="br0">&#40;</span>skills<span class="sy0">.</span>abilSpawnMinions<span class="sy0">:</span>GetCharges<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">==</span>skills<span class="sy0">.</span>abilSpawnMinions<span class="sy0">:</span>GetLevel<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">*</span><span class="nu0">3</span> <span class="kw2">and</span> unitSelf<span class="sy0">:</span>GetHealthPercent<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">*</span><span class="nu0">1000</span><span class="sy0">==</span><span class="nu0">1000</span><span class="br0">&#41;</span><span class="br0">&#41;</span><span class="kw1">then</span> <span class="kw1">return</span> <span class="nu0">0</span> <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> closestCorpse <span class="sy0">=</span> <span class="kw4">nil</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nClosestCorpseDistSq <span class="sy0">=</span> <span class="nu0">9999</span><span class="sy0">*</span><span class="nu0">9999</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">for</span> key<span class="sy0">,</span> v <span class="kw2">in</span> <span class="kw3">pairs</span><span class="br0">&#40;</span>object<span class="sy0">.</span>tCorpseTable<span class="br0">&#41;</span> <span class="kw1">do</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; vecCorpsePosition <span class="sy0">=</span> v<span class="br0">&#91;</span><span class="st0">'position'</span><span class="br0">&#93;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="co1">--&quot;safe&quot; corpses aren't toward the opponents.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="kw2">not</span> object<span class="sy0">.</span>vecLaneForward <span class="kw2">or</span> <span class="kw3">abs</span><span class="br0">&#40;</span>core<span class="sy0">.</span>RadToDeg<span class="br0">&#40;</span>core<span class="sy0">.</span>AngleBetween<span class="br0">&#40;</span>vecCorpsePosition <span class="sy0">-</span> unitSelf<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="sy0">,</span> <span class="sy0">-</span>object<span class="sy0">.</span>vecLaneForward<span class="br0">&#41;</span><span class="br0">&#41;</span> <span class="br0">&#41;</span> <span class="sy0">&lt;</span> <span class="nu0">130</span> <span class="kw1">then</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">local</span> nDistSq <span class="sy0">=</span> Vector3<span class="sy0">.</span>Distance2DSq<span class="br0">&#40;</span>vecCorpsePosition<span class="sy0">,</span> unitSelf<span class="sy0">:</span>GetPosition<span class="br0">&#40;</span><span class="br0">&#41;</span><span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>bDebugLines<span class="br0">&#41;</span><span class="kw1">then</span> core<span class="sy0">.</span>DrawXPosition<span class="br0">&#40;</span>vecCorpsePosition<span class="sy0">,</span> <span class="st0">'aqua'</span><span class="br0">&#41;</span> <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> nDistSq <span class="sy0">&lt;</span> nClosestCorpseDistSq <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; closestCorpse <span class="sy0">=</span> v</div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; nClosestCorpseDistSq <span class="sy0">=</span> nDistSq</div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">else</span> <span class="kw1">if</span> <span class="br0">&#40;</span>bDebugLines<span class="br0">&#41;</span><span class="kw1">then</span> core<span class="sy0">.</span>DrawXPosition<span class="br0">&#40;</span>vecCorpsePosition<span class="sy0">,</span> <span class="st0">'black'</span><span class="br0">&#41;</span> <span class="kw1">end</span> <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>closestCorpse<span class="br0">&#41;</span> <span class="kw1">then</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; corpsePosition<span class="sy0">=</span>closestCorpse<span class="br0">&#91;</span><span class="st0">'position'</span><span class="br0">&#93;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> <span class="nu0">30</span><span class="sy0">*</span><span class="br0">&#40;</span><span class="br0">&#40;</span><span class="nu0">500</span><span class="sy0">-</span><span class="kw3">sqrt</span><span class="br0">&#40;</span>nClosestCorpseDistSq<span class="br0">&#41;</span><span class="br0">&#41;</span><span class="sy0">/</span><span class="nu0">500</span><span class="br0">&#41;</span> <span class="co1">-- basically, the closer you are to the corpse, the more you want it. 30 utility over 500 units.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">else</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; corpsePosition<span class="sy0">=</span><span class="kw4">nil</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> <span class="nu0">0</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1"><span class="kw1">local</span> <span class="kw1">function</span> getCorpsesExecute<span class="br0">&#40;</span>botBrain<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">if</span> <span class="br0">&#40;</span>corpsePosition<span class="br0">&#41;</span> <span class="kw1">then</span> <span class="co1">--corpse exists</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> core<span class="sy0">.</span>OrderMoveToPosClamp<span class="br0">&#40;</span>botBrain<span class="sy0">,</span>core<span class="sy0">.</span>unitSelf<span class="sy0">,</span> corpsePosition<span class="br0">&#41;</span><span class="co1">--get the corpse.</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="kw1">return</span> <span class="kw4">false</span></div></li>
-<li class="li2"><div class="de2"><span class="kw1">end</span></div></li>
-<li class="li1"><div class="de1">behaviorLib<span class="sy0">.</span>getCorpses <span class="sy0">=</span> <span class="br0">&#123;</span><span class="br0">&#125;</span></div></li>
-<li class="li2"><div class="de2">behaviorLib<span class="sy0">.</span>getCorpses<span class="br0">&#91;</span><span class="st0">&quot;Utility&quot;</span><span class="br0">&#93;</span> <span class="sy0">=</span> getCorpsesUtility</div></li>
-<li class="li1"><div class="de1">behaviorLib<span class="sy0">.</span>getCorpses<span class="br0">&#91;</span><span class="st0">&quot;Execute&quot;</span><span class="br0">&#93;</span> <span class="sy0">=</span> getCorpsesExecute</div></li>
-<li class="li2"><div class="de2">behaviorLib<span class="sy0">.</span>getCorpses<span class="br0">&#91;</span><span class="st0">&quot;Name&quot;</span><span class="br0">&#93;</span> <span class="sy0">=</span> <span class="st0">&quot;getCorpses&quot;</span></div></li>
-<li class="li1"><div class="de1"><span class="kw3">tinsert</span><span class="br0">&#40;</span>behaviorLib<span class="sy0">.</span>tBehaviors<span class="sy0">,</span> behaviorLib<span class="sy0">.</span>getCorpses<span class="br0">&#41;</span></div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1">&nbsp;<span class="co1">--@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ PERSONALITY</span></div></li>
-<li class="li2"><div class="de2">core<span class="sy0">.</span>tKillChatKeys<span class="sy0">=</span><span class="br0">&#123;</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="st0">&quot;*Spews all over keyboard*&quot;</span><span class="sy0">,</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="st0">&quot;Get trashed kid.&quot;</span><span class="sy0">,</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="st0">&quot;BWAHAHAHAHA!&quot;</span><span class="sy0">,</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="st0">&quot;Yummy!&quot;</span><span class="sy0">,</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="st0">&quot;Feeling.. trapped?&quot;</span></div></li>
-<li class="li2"><div class="de2"><span class="br0">&#125;</span></div></li>
-<li class="li1"><div class="de1">core<span class="sy0">.</span>tDeathChatKeys <span class="sy0">=</span> <span class="br0">&#123;</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="st0">&quot;Going to leave me here to rot?&quot;</span><span class="sy0">,</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="st0">&quot;Bargh, my minions got stuck..&quot;</span><span class="sy0">,</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="st0">&quot;You make me sick!&quot;</span><span class="sy0">,</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="st0">&quot;I got S2'd..&quot;</span><span class="sy0">,</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="st0">&quot;Team? TEAM??&quot;</span><span class="sy0">,</span></div></li>
-<li class="li1"><div class="de1">&nbsp; &nbsp; &nbsp; &nbsp; <span class="st0">&quot;Hey, you used a cheat code!&quot;</span><span class="sy0">,</span></div></li>
-<li class="li2"><div class="de2">&nbsp; &nbsp; &nbsp; &nbsp; <span class="st0">&quot;Go my minions!&quot;</span></div></li>
-<li class="li1"><div class="de1"><span class="br0">&#125;</span> </div></li>
-<li class="li2"><div class="de2">&nbsp;</div></li>
-<li class="li1"><div class="de1">&nbsp;<span class="co1">--@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ SIGN OFF</span></div></li>
-<li class="li2"><div class="de2">BotEcho<span class="br0">&#40;</span><span class="st0">'finished loading balfagore_main'</span><span class="br0">&#41;</span></div></li>
-</ol></div>
-		</div>
-	</div></div>
-	<div class="content_title">
-		<span class="span_right raw_links"><a href="/index/w7fj3tgY" rel="nofollow">create a <u>new version</u> of this paste</a></span>
-		RAW Paste Data
-	</div>
-	<form class="paste_form" id="myform" method="post" action="/post.php">
-		<div class="textarea_border">
-			<textarea id="paste_code" class="paste_code" name="paste_code" onkeydown="return catchTab(this,event)">----------------------------------------------
--- _____       
---|_   _|
---  | |   _ __   __ _  _ __    __ _   __ _   ___   _ __   ___
---  | |  | '__| / _` || '_ \  / _` | / _` | / _ \ | '__| / _ \
---  | |  | |   | (_| || |_) || (_| || (_| || (_) || |   |  __/
---  \_/  |_|    \__,_|| .__/  \__,_| \__, | \___/ |_|    \___|
---                    | |             __/ |
---                    |_|            |___/     version 0.8.3
---
-------------------------------------------
---         Created by: kairus101        --
-------------------------------------------
-
-------------------------------------------
---          Bot Initialization          --
-------------------------------------------
-local _G = getfenv(0)
-local object = _G.object
-object.myName = object:GetName()
-object.bRunLogic,object.bRunBehaviors,object.bUpdates,object.bUseShop,object.bRunCommands,object.bMoveCommands,object.bAttackCommands,object.bAbilityCommands,object.bOtherCommands=true
-object.logger = {}
-object.bReportBehavior,object.bDebugUtility,object.logger.bWriteLog,object.logger.bVerboseLog=false
-object.core 		= {}
-object.eventsLib 	= {}
-object.metadata 	= {}
-object.behaviorLib 	= {}
-object.skills 		= {}
-runfile &quot;bots/core.lua&quot;
-runfile &quot;bots/botbraincore.lua&quot;
-runfile &quot;bots/eventslib.lua&quot;
-runfile &quot;bots/metadata.lua&quot;
-runfile &quot;bots/behaviorlib.lua&quot;
-local core, eventsLib, behaviorLib, metadata, skills = object.core, object.eventsLib, object.behaviorLib, object.metadata, object.skills
-local print, ipairs, pairs, string, table, next, type, tinsert, tremove, tsort, format, tostring, tonumber, strfind, strsub	= _G.print, _G.ipairs, _G.pairs, _G.string, _G.table, _G.next, _G.type, _G.table.insert, _G.table.remove, _G.table.sort, _G.string.format, _G.tostring, _G.tonumber, _G.string.find, _G.string.sub
-local ceil, floor, pi, tan, atan, atan2, abs, cos, sin, acos, max, random, sqrt	= _G.math.ceil, _G.math.floor, _G.math.pi, _G.math.tan, _G.math.atan, _G.math.atan2, _G.math.abs, _G.math.cos, _G.math.sin, _G.math.acos, _G.math.max, _G.math.random, _G.math.sqrt
-local BotEcho, VerboseLog, BotLog = core.BotEcho, core.VerboseLog, core.BotLog
-local Clamp = core.Clamp
-BotEcho('loading balfagore_main...')
-
----------------------------------
---          Constants          --
----------------------------------
---Balphagore
-object.heroName = 'Hero_Bephelgor'
-
--- Item buy order. internal names
-behaviorLib.StartingItems = {&quot;Item_LoggersHatchet&quot;, &quot;Item_GuardianRing&quot;, &quot;2 Item_RunesOfTheBlight&quot;}
-behaviorLib.LaneItems = {&quot;Item_ManaRegen3&quot;,&quot;Item_Marchers&quot;,&quot;Item_MysticVestments&quot;,&quot;Item_EnhancedMarchers&quot;}
-behaviorLib.MidItems = {&quot;Item_Energizer&quot;,&quot;Item_Pierce 3&quot;, &quot;Item_SolsBulwark&quot;, &quot;Item_DaemonicBreastplate&quot;}
-behaviorLib.LateItems = {&quot;Item_BehemothsHeart&quot;, 'Item_Damage9'}
-
--- Skillbuild. 0 is Regurgitate, 1 is Demonic Pathogen, 2 is Corpse Conversion, 3 is Hell on Newerth, 4 is Attributes
-object.tSkills = {
-	2, 0, 2, 0, 2, 0,	-- 1-6
-	2, 0, 3, 4, 3, 		-- 7-11
-	4, 4, 4, 4, 3,		-- 12-16
-	4, 4, 4, 4, 4, 1, 1, 1, 1,	--17-25
-}
-
--- Bonus agression points if a skill/item is available for use
-object.nHellUp = 5
-object.nRegurgitateUp = 10
-object.nMinionsCloseUp = 8
-object.nTrappingUp = 10
-object.nEnergizerUp = 8
-object.nCanAttackUp = 10
-
--- Bonus agression points that are applied to the bot upon successfully using a skill/item
-object.nRegurgitateUse = 10
-object.nHellUse = 8
-object.nEnergizerUse = 8
-
--- Thresholds of aggression the bot must reach to use these abilities
-object.nRegurgitateThreshold = 26
-object.tHellOnNewerthThreshold = {70,60,50} --depending on charges
-object.nEnergizerThreshold = 30
-
---time variables
-object.nTimeBarfed=0 --used to get corpses at well
-object.nTimeEnergizered=0 --used for minions to remember when to trap again
-
---table variables (need to keep track on minions and corpses)
-object.tMinions={} --hold minions
-object.tCorpseTable={}--position,time
-object.tOldLocalCreeps={}--for checking if units are dead for corpses.
-
---item variables
-core.itemEnergizer=nil --energizer
-core.itemBols=nil --sol's bolwork
-
---other variables
-core.distSqTolerance=5*5
-behaviorLib.nCreepPushbackMul = 0.4 --0.5
-behaviorLib.nTargetPositioningMul = 0.4 --0.6
-behaviorLib.nTargetCriticalPositioningMul = 1 --2
-object.nRealDistance=0 --average distance minions are from target.
-object.nMinionsClose=0 --this is for use with energizer
-object.bUnitStill=false --this adds extra harass utility if hero is trapped
-object.nMinionSkip=5 --control minions every object.nMinionSkip frames.
-object.nCurSkip=0 --current frame. This is for frame skip.
-
---difficulty variables
---[[
-core.nEASY_DIFFICULTY 	= 1
-core.nMEDIUM_DIFFICULTY = 2
-core.nHARD_DIFFICULTY 	= 3
-]]
-object.trapEffectiveness={95*95,75*75,55*55}--Lower is better.
-object.trapCycleSkip={7,4,1}--Lower is better.
-object.trapCircleRadius={550,450,350}
-
-------------------------------
---          Skills          --
-------------------------------
-function object:SkillBuild()
-	local unitSelf = self.core.unitSelf
-	if  skills.abilRegurgitate == nil then
-		skills.abilRegurgitate = unitSelf:GetAbility(0)--Barf
-		skills.abilDemonicPathogen = unitSelf:GetAbility(1)--Useless silence
-		skills.abilSpawnMinions = unitSelf:GetAbility(2)--Corpse pickup/Minions
-		skills.abilHellOnNewerth = unitSelf:GetAbility(3)--Ultimate
-		skills.abilAttributeBoost = unitSelf:GetAbility(4)--Stats
-	end
-	if unitSelf:GetAbilityPointsAvailable() &lt;= 0 then
-		return
-	end
-	local nlev = unitSelf:GetLevel()
-	local nlevpts = unitSelf:GetAbilityPointsAvailable()
-	for i = nlev, nlev+nlevpts do
-		unitSelf:GetAbility( object.tSkills[i] ):LevelUp()
-	end
-end
-
-------------------------------------------
---          FindItems Override          --
-------------------------------------------
-local function FindItemsOverride(botBrain)
-	local bUpdated = object.FindItemsOld(botBrain)
-
-	if core.itemEnergizer ~= nil and not core.itemEnergizer:IsValid() then
-		core.itemEnergizer = nil
-	end
-	if core.itemGhostMarchers ~= nil and not core.itemGhostMarchers:IsValid() then
-		core.itemGhostMarchers = nil
-	end
-
-	if bUpdated then
-		if core.itemEnergizer and core.itemGhostMarchers then
-			return
-		end
-
-		local inventory = core.unitSelf:GetInventory(false)
-		for slot = 1, 6 do
-			local curItem = inventory[slot]
-			if curItem then
-				if core.itemEnergizer == nil and curItem:GetName() == &quot;Item_Energizer&quot; then
-					core.itemEnergizer = core.WrapInTable(curItem)
-				elseif core.itemGhostMarchers == nil and curItem:GetName() == &quot;Item_EnhancedMarchers&quot; then
-					core.itemGhostMarchers = core.WrapInTable(curItem)
-				elseif core.itemBols == nil and curItem:GetName() == &quot;Item_SolsBulwark&quot; then
-					core.itemBols = core.WrapInTable(curItem)
-					core.OrderItemClamp(botBrain, unitSelf, core.itemBols) --turn to negative bols
-				end
-			end
-		end
-	end
-end
-object.FindItemsOld=core.FindItems
-core.FindItems = FindItemsOverride
-
-----------------------------------------------------
---        Minion Trapping Helper Functions        --
-----------------------------------------------------
-local function GetHeroByID(botBrain, ID) --refreshes the targetted unit.
-	for id, unitHero in pairs(HoN.GetHeroes(core.enemyTeam)) do
-		if unitHero ~= nil then
-			if core.CanSeeUnit(botBrain, unitHero) then
-				if unitHero:GetUniqueID()==ID then
-					return unitHero
-				end
-			end
-		end
-	end
-	return nil
-end
-local function positionOffset(pos, angle, distance) --this is used by minions to form a ring around people.
-	tmp = Vector3.Create(cos(angle)*distance,sin(angle)*distance)
-	return tmp+pos
-end
-local function vectorLength(vec) --easy way to get the distance between two points. Much more efficient than pythagoras' theorem - thanks to Anakonda.
-	local x = vec.x
-	local y = vec.y
-	if x==0 then return y end
-	if y==0 then return x end
-	local angle = atan(y/x)
-	return y/sin(angle)
-end
-
-----------------------------------------
---          OnThink Override          --
-----------------------------------------
-function object:onthinkOverride(tGameVariables) --This is run, even while dead. Every frame.
-	self:onthinkOld(tGameVariables)--don't distrupt old think, run it.
-
-	local unitSelf = core.unitSelf
-	local vecSelfPos = unitSelf:GetPosition()
-	local nNumMinions=table.getn(object.tMinions)
-	local bDebugLines = false
-	
-	---------------------------------
-	--      Register Minions       --
-	---------------------------------
-	if (nNumMinions==0 and unitSelf:IsAlive() and core.localUnits~=nil)then
-		object.tMinions={}
-		i=1--keep track of array index
-		for key, unit in pairs(core.localUnits[&quot;AllyUnits&quot;]) do
-			if (unit:GetTypeName()==&quot;Pet_Bephelgor_Ability2&quot; and unit:GetOwnerPlayer()==unitSelf:GetOwnerPlayer()) then
-				object.tMinions[i]=unit
-				if (bDebugLines) then core.DrawDebugArrow(object.tMinions[i]:GetPosition(), vecSelfPos, 'white') end
-				i=i+1
-			end
-		end
-	end
-	
-	-----------------------------------------------------------------------------------
-	---------------------------------
-	--       Control Minions       --
-	---------------------------------
-	-- This part is the most complicated part of the program. The minions:
-	-- 1. Find a hero or building.
-	-- 2. if target is a building, attack it.
-	-- 3. If it's a hero:
-	--   a. If target/minion is unit walking, just attack them.
-	--   b. If target is not trapped yet and we have &gt;3 minions, surround it,
-	--        and close in until they are trapped.
-	--   c. Attack trapped target.
-	-- 4. If there is no target, attack-move along with the creep wave.
-	-----------------------------------------------------------------------------------
-	
-	--this is a frame-skip. We don't need to run minions code every single frame.
-	object.nCurSkip=object.nCurSkip+1
-	if (object.nCurSkip&gt;=object.nMinionSkip and table.getn(object.tMinions)&gt;0) then--don't skip
-		object.nCurSkip=0
-		local botBrain=self
-		local bDebugLines = false
-		local vecWellPos=(core.enemyWell and core.enemyWell:GetPosition())
-		
-		---------------------------------
-		--    1.Minions Get Target     --
-		---------------------------------		
-		local target=nil --this is what minions will use
-		local heroTarget=behaviorLib.heroTarget
-		local buildingTarget=behaviorLib.hitBuildingTarget
-		if (heroTarget and heroTarget:IsValid() and heroTarget:IsAlive() and heroTarget:GetHealth()&gt;0 and core.CanSeeUnit(botBrain, heroTarget) and Vector3.Distance2DSq(heroTarget:GetPosition(), vecSelfPos)&lt;2000*2000) then
-			--If we have a hero in mind
-			target=GetHeroByID(botBrain, heroTarget:GetUniqueID())
-		elseif (buildingTarget and buildingTarget:IsValid() and buildingTarget:IsAlive() and core.CanSeeUnit(botBrain, buildingTarget) and Vector3.Distance2DSq(buildingTarget:GetPosition(), vecSelfPos)&lt;2000*2000) then
-			--If we have a building in mind
-			target=buildingTarget
-		end
-		object.nRealDistance=nil
-		object.nMinionsClose=0
-		
-		if (target and bDebugLines) then core.DrawDebugArrow(core.unitSelf:GetPosition(), target:GetPosition(), 'white') end
-		
-		---------------------------------
-		--  2.Minions Building Attack  --
-		---------------------------------
-		if (target and target:IsBuilding()) then --kill buildings trapagore attacks, first and foremost.
-			for i=1,nNumMinions do
-				if (object.tMinions[i] and object.tMinions[i]:IsValid() and object.tMinions[i]:IsAlive()) then
-					core.OrderAttack(botBrain, object.tMinions[i], target,false)
-				end
-			end
-			object.nMinionSkip=10--500ms till next cycle
-			
-		---------------------------------
-		--    3.Minions Hero Attack    --
-		---------------------------------
-		elseif (target and nNumMinions&gt;2)then
-			local target=target
-			local vecTargetPos=target:GetPosition()--+vTargetVel*0.5
-			local angleToWell= atan2(vecWellPos.y-vecTargetPos.y,vecWellPos.x-vecTargetPos.x)
-			
-			if (nNumMinions&gt;3) then
-				object.nMinionSkip=object.trapCycleSkip[core.nDifficulty]
-				
-				local totalDistance=0
-				for i=1,nNumMinions do
-					if (object.tMinions[i] and object.tMinions[i]:IsValid() and object.tMinions[i]:IsAlive() and object.tMinions[i]:GetPosition() and target and vecTargetPos) then
-						vecMinionPos=object.tMinions[i]:GetPosition()
-						totalDistance=totalDistance+vectorLength(vecMinionPos-vecTargetPos)
-					end
-				end
-				--local nAvgDistance=1.1*sqrt(tTotalDisplacement[1]*tTotalDisplacement[1]+tTotalDisplacement[2]*tTotalDisplacement[2])/nNumMinions
-				local nAvgDistance=totalDistance/nNumMinions
-				
-				object.nRealDistance=nAvgDistance
-				distance=nAvgDistance-30
-				if (distance&gt;object.trapCircleRadius[core.nDifficulty])then distance=object.trapCircleRadius[core.nDifficulty] end --There is a cap on the distance.
-				
-				--TRAP IF POSSIBLE
-				i=1
-				while (i&lt;=nNumMinions) do
-					if (object.tMinions[i] and object.tMinions[i]:IsValid() and object.tMinions[i]:IsAlive() and object.tMinions[i]:GetPosition()) then
-						vecTrappingPosition = positionOffset(vecTargetPos, angleToWell +2*pi*i/nNumMinions, distance)
-						nDistFromTargetSq=Vector3.Distance2DSq(object.tMinions[i]:GetPosition(), vecTargetPos)
-						-- Don't try to trap if you or they are unitwalking, or trapping already.
-						if (target:GetUnitwalking() or HoN:GetGameTime()-object.nTimeEnergizered&lt;5200 or (HoN:GetGameTime()-object.nTimeEnergizered&gt;6000 and object.tMinions[i]:GetUnitwalking()) or target:GetHealthPercent()*100&lt;5 or (Vector3.Distance2DSq(object.tMinions[i]:GetPosition(), vecTrappingPosition)&lt;object.trapEffectiveness[core.nDifficulty]) and nDistFromTargetSq&lt;object.trapEffectiveness[core.nDifficulty]) then--35 was tmp
-							---------------------------------
-							--       a.Attack Target       --
-							---------------------------------
-							if (nDistFromTargetSq&gt;128*128) then
-								--ATTACK, You are/can't trap them.
-								if (core.GetAttackSequenceProgress(object.tMinions[i])~=&quot;windup&quot;) then--this is needed, don't cancel attacks param doesn't work for some reason.
-									core.OrderMoveToPos(botBrain,object.tMinions[i], vecTargetPos,false)
-								end
-								if (bDebugLines)then core.DrawXPosition(vecTargetPos, 'yellow') end
-							else
-								core.OrderAttack(botBrain, object.tMinions[i], target,false)
-								if (bDebugLines)then core.DrawXPosition(vecTargetPos, 'red') end
-							end
-						else --TRAP, they are trappable.
-							---------------------------------
-							--      b.Surround Target      --
-							---------------------------------
-							botBrain:OrderPosition(object.tMinions[i].object or object.tMinions[i], &quot;Move&quot;,vecTrappingPosition,nil, nil,false)
-							if (bDebugLines)then core.DrawDebugArrow(object.tMinions[i]:GetPosition(), vecTrappingPosition, 'green') end--draw where hero is going
-						end
-						i=i+1
-					else
-						table.remove(object.tMinions,i)
-						nNumMinions=table.getn(object.tMinions)
-					end
-				end
-			else
-				---------------------------------
-				--       c.Attack  Target      --
-				---------------------------------
-				--we don't have enough minions to trap, just attack.
-				object.nMinionSkip=5--250ms till next cycle
-				for i=1,nNumMinions do
-					if (object.tMinions[i] and object.tMinions[i]:IsValid() and object.tMinions[i]:IsAlive()) then core.OrderAttack(botBrain, object.tMinions[i], target) end --kill em'
-				end
-			end
-			
-		---------------------------------
-		--          4.Pushing          --
-		---------------------------------
-		else --no target, push forwards.
-			object.nMinionSkip=5--250ms till next cycle
-			destPos=core.GetFurthestCreepWavePos(core.tMyLane, true)
-			for i=1,nNumMinions do
-				if (object.tMinions[i])then
-					if (object.tMinions[i] and object.tMinions[i]:IsValid() and object.tMinions[i]:IsAlive()) then
-						core.OrderAttackPosition(botBrain, object.tMinions[i], destPos,false,false)--attackmove
-					else table.remove(object.tMinions,i) i=i-1 nNumMinions=table.getn(object.tMinions) end
-				end
-			end
-		end
-		nNumMinions=table.getn(object.tMinions)
-		
-	end
-end
-object.onthinkOld = object.onthink
-object.onthink 	= object.onthinkOverride
-
-----------------------------------------------
---          OnCombatEvent Override          --
-----------------------------------------------
-function object:oncombateventOverride(EventData)
-	self:oncombateventOld(EventData)
-
-	local nAddBonus = 0
-
-	if EventData.Type == &quot;Ability&quot; then
-		if EventData.InflictorName == &quot;Ability_Bephelgor1&quot; then
-			nAddBonus = nAddBonus + object.nRegurgitateUse
-		elseif EventData.InflictorName == &quot;Ability_Bephelgor4&quot; then
-			nAddBonus = nAddBonus + object.nHellUse
-		end
-	elseif EventData.Type == &quot;Item&quot; then
-		if core.itemEnergizer ~= nil and EventData.SourceUnit == core.unitSelf:GetUniqueID() and EventData.InflictorName == core.itemEnergizer:GetName() then
-			nAddBonus = nAddBonus + self.nEnergizerUse
-		end
-	end
-
-	if nAddBonus &gt; 0 then
-		core.DecayBonus(self)
-		core.nHarassBonus = core.nHarassBonus + nAddBonus
-	end
-end
-object.oncombateventOld = object.oncombatevent
-object.oncombatevent = object.oncombateventOverride
-
-----------------------------------------------------
---          CustomHarassUtility Override          --
-----------------------------------------------------
-local function CustomHarassUtilityOverride(unit) --this is part of harass hero utility.
-	unitSelf=core.unitSelf
-	local val = 0
-	local heroTarget = unit
-	--can attack enemy
-	if (Vector3.Distance2DSq(unitSelf:GetPosition(), heroTarget:GetPosition())&lt;128*128 and unitSelf:IsAttackReady()) then val=val+object.nCanAttackUp end
-	--can use ultimate.
-	if skills.abilHellOnNewerth:CanActivate() then val = val + object.nHellUp end
-	--can use barf
-	if skills.abilRegurgitate:CanActivate() then val = val + object.nRegurgitateUp end
-	local nNumMinions=#object.tMinions --this is the same as table.getn(minions)
-	--Minions are close to enemy
-	if (nNumMinions&gt;3 and object.nRealDistance and object.nRealDistance&lt;300) then val=val+object.nMinionsCloseUp end
-	--Minions are trapping enemy
-	if (nNumMinions&gt;3 and object.bUnitStill and object.nRealDistance and object.nRealDistance&lt;80)then val=val+object.nTrappingUp end--trapping! 10
-	--can use energizer
-	if (core.itemEnergizer and object.nMinionsClose&gt;3 and core.itemEnergizer:CanActivate()) then val=val+object.nEnergizerUp end --can activate energizer
-	return val
-end
-behaviorLib.CustomHarassUtility = CustomHarassUtilityOverride
-
-----------------------------------------------------
---             Last hitting Overrides             --
-----------------------------------------------------
-local function KaiGetCreepAttackTargetOverride(botBrain, unitEnemyCreep, unitAllyCreep)
-	local unitSelf = core.unitSelf
-	
-	--Get closest hero. We can make decisions based on this.
-	local uClosestHero = nil
-	local nClosestHeroDistSq = 2000*2000 -- more than 2000 and we are not concerned. 2000*2000
-	for id, unitHero in pairs(HoN.GetHeroes(core.enemyTeam)) do
-		if unitHero ~= nil then
-			if core.CanSeeUnit(botBrain, unitHero) and unitHero:GetTeam()~=team then
-				local nDistanceSq = Vector3.Distance2DSq(unitHero:GetPosition(), core.unitSelf:GetPosition())
-				if nDistanceSq &lt; nClosestHeroDistSq then
-					nClosestHeroDistSq = nDistanceSq
-					uClosestHero = unitHero
-				end
-			end
-		end
-	end
-	
-	--addition is how much sooner to move towards creep. '30' is 30hp more than you can 1 hit kill it for.
-	addition=30
-	if (uClosestHero and uClosestHero:GetAttackType() ~= &quot;melee&quot;) then --be more passive if opponent is ranged.
-		addition=20
-	end	
-	
-	if (uClosestHero and unitSelf:GetHealthPercent()*100&lt;60 and Vector3.Distance2D(unitSelf:GetPosition(), uClosestHero:GetPosition())&lt;uClosestHero:GetAttackRange()+150) then return nil end --too dangerous. Stay back.
-	local nDamageAverage = unitSelf:GetFinalAttackDamageMin()+addition --make the hero go to the unit when it is 40 hp away, but too risky if enemy is ranged
-	core.FindItems(botBrain)
-	if core.itemHatchet then
-		nDamageAverage = nDamageAverage * core.itemHatchet.creepDamageMul
-	end
-	if core.nDifficulty == core.nEASY_DIFFICULTY then
-		nDamageAverage = nDamageAverage + 10
-	end
-	if unitEnemyCreep and core.CanSeeUnit(botBrain, unitEnemyCreep) then --return enemy creep
-		local nTargetHealth = unitEnemyCreep:GetHealth()
-		if nDamageAverage &gt;= nTargetHealth or core.teamBotBrain.nPushState==2 then
-			return unitEnemyCreep
-		end
-	end
-	if unitAllyCreep and core.nDifficulty ~= core.nEASY_DIFFICULTY then
-		local nTargetHealth = unitAllyCreep:GetHealth()
-		if nDamageAverage &gt;= nTargetHealth then
-			if core.teamBotBrain.nPushState~=2 and (core.nDifficulty ~= core.nEASY_DIFFICULTY or (core.bIsTutorial and core.bTutorialBehaviorReset == true and core.myTeam == HoN.GetHellbourneTeam())) then
-				return unitAllyCreep
-			end
-		end
-	end
-	return nil
-end
-object.GetCreepAttackTargetOld = behaviorLib.GetCreepAttackTarget
-behaviorLib.GetCreepAttackTarget = KaiGetCreepAttackTargetOverride
-
-local function KaiAttackCreepsExecuteOverride(botBrain)
-	local unitSelf = core.unitSelf
-	local currentTarget = core.unitCreepTarget
-	if currentTarget and core.CanSeeUnit(botBrain, currentTarget) then
-		local vecTargetPos = currentTarget:GetPosition()
-		local nDistSq = Vector3.Distance2DSq(unitSelf:GetPosition(), vecTargetPos)
-		local nAttackRangeSq = core.GetAbsoluteAttackRangeToUnit(unitSelf, currentTarget, true)
-		local nDamageAverage = unitSelf:GetFinalAttackDamageMin()
-		if core.itemHatchet then
-			nDamageAverage = nDamageAverage * core.itemHatchet.creepDamageMul
-		end
-		if currentTarget ~= nil then			
-			if nDistSq &lt; nAttackRangeSq and unitSelf:IsAttackReady() and (nDamageAverage*(1-currentTarget:GetPhysicalResistance())&gt;=currentTarget:GetHealth() or core.teamBotBrain.nPushState==2) then --only kill if you can get gold
-				return core.OrderAttackClamp(botBrain, unitSelf, currentTarget)
-			elseif nDistSq &gt; nAttackRangeSq and vecTargetPos and unitSelf:IsAttackReady() and (nDamageAverage*(1-currentTarget:GetPhysicalResistance())&gt;=currentTarget:GetHealth()-100 or core.teamBotBrain.nPushState==2) then
-				return core.OrderMoveToPosClamp(botBrain, unitSelf, vecTargetPos) --moves hero to target
-			else
-				return core.OrderHoldClamp(botBrain, unitSelf, false) --this is where the magic happens. Wait for the kill.
-			end
-		end
-	end
-	return false
-end
-object.AttackCreepsExecuteOld = behaviorLib.AttackCreepsBehavior[&quot;Execute&quot;]
-behaviorLib.AttackCreepsBehavior[&quot;Execute&quot;] = KaiAttackCreepsExecuteOverride
-
-----------------------------------------------------
---             Heal At Well Override              --
-----------------------------------------------------
---return to well more often. --2000 gold adds 8 to return utility, 0% mana also adds 8.
---When returning to well, use skills and items.
-local function HealAtWellUtilityOverride(botBrain)
-	if (Vector3.Distance2DSq(core.unitSelf:GetPosition(), core.allyWell and core.allyWell:GetPosition() or behaviorLib.PositionSelfBackUp())&lt;400*400 and core.unitSelf:GetManaPercent()*100&lt;95) then return 80 end
-	return object.HealAtWellUtilityOld(botBrain)*1.75+(botBrain:GetGold()*8/2000)+ 8-(core.unitSelf:GetManaPercent()*8) --couragously flee back to base.
-end
-local function HealAtWellExecuteOverride(botBrain)
-	local vecWellPos = core.allyWell and core.allyWell:GetPosition() or behaviorLib.PositionSelfBackUp()
-	if (Vector3.Distance2DSq(core.unitSelf:GetPosition(), vecWellPos)&gt;600*600)then
-		if (core.itemEnergizer and core.itemEnergizer:CanActivate() and not unitSelf:HasState(&quot;State_Energizer_Buff&quot;))then --when heading to base, use energizer
-			botBrain:OrderItem(core.itemEnergizer.object or core.itemEnergizer, false)
-			object.nTimeEnergizered=HoN:GetGameTime()
-		end
-		if (core.itemGhostMarchers and core.itemGhostMarchers:CanActivate())then --when heading to base, use boots
-			botBrain:OrderItem(core.itemGhostMarchers.object or core.itemGhostMarchers, false)
-		end
-		if (skills.abilRegurgitate:CanActivate()) then --BARF
-			core.OrderAbility(botBrain, skills.abilRegurgitate)
-		end
-	end
-	return object.HealAtWellExecuteOld(botBrain)
-end
-object.HealAtWellUtilityOld = behaviorLib.HealAtWellBehavior[&quot;Utility&quot;]
-object.HealAtWellExecuteOld = behaviorLib.HealAtWellBehavior[&quot;Execute&quot;]
-behaviorLib.HealAtWellBehavior[&quot;Utility&quot;] = HealAtWellUtilityOverride
-behaviorLib.HealAtWellBehavior[&quot;Execute&quot;] = HealAtWellExecuteOverride
-
-----------------------------------------
---          Harass Behaviour          --
-----------------------------------------
-local function HarassHeroExecuteOverride(botBrain)
-	local unitSelf = core.unitSelf
-	local nHarassUtility = behaviorLib.lastHarassUtil
-	local unitTarget = behaviorLib.heroTarget
-	
-	local nTargetDistanceSq = Vector3.Distance2DSq(unitSelf:GetPosition(), unitTarget:GetPosition())
-	local nNumMinions=table.getn(object.tMinions)
-	local vecWellPos = core.allyWell and core.allyWell:GetPosition() or behaviorLib.PositionSelfBackUp()
-	local distToWellSq=Vector3.Distance2DSq(unitSelf:GetPosition(), vecWellPos)
-	local bActionTaken = false
-	
-	--Spawn Minions
-	if skills.abilSpawnMinions:CanActivate() and distToWellSq&gt;500*500 and unitSelf:GetLevel()&gt;=5 and skills.abilSpawnMinions:GetCharges()==skills.abilSpawnMinions:GetLevel()*3 and nNumMinions&lt;3 and --can activate
-		(core.teamBotBrain.nPushState==2 or behaviorLib.heroTarget) then
-		bActionTaken=core.OrderAbility(botBrain,skills.abilSpawnMinions)
-	end
-	
-	--Regurgitate
-	if not bActionTaken and nHarassUtility&gt;object.nRegurgitateThreshold and nTargetDistanceSq&lt;300*300 and skills.abilRegurgitate:CanActivate() then
-		object.nTimeBarfed=HoN:GetGameTime()
-		bActionTaken = core.OrderAbility(botBrain, skills.abilRegurgitate,false,false)
-	end
-	
-	-- Ultimate
-	local ultThreshholdValue=floor(skills.abilHellOnNewerth:GetCharges()/40)+1
-	if not bActionTaken and skills.abilHellOnNewerth:CanActivate() and nHarassUtility&gt;object.tHellOnNewerthThreshold[ultThreshholdValue] and nTargetDistanceSq&lt;550*550 then
-		bActionTaken = core.OrderAbility(botBrain, skills.abilHellOnNewerth)
-	end
-	
-	-- Energizer
-	if not bActionTaken and core.itemEnergizer and core.itemEnergizer:CanActivate() and nHarassUtility&gt;object.nEnergizerThreshold then
-		botBrain:OrderItem(core.itemEnergizer.object or core.itemEnergizer, false)
-		object.nTimeEnergizered=HoN:GetGameTime()
-	end
-	
-	--default
-	if not bActionTaken then
-		return object.harassHeroExecuteOld(botBrain)
-	end
-	return bActionTaken
-end
-object.harassHeroExecuteOld = behaviorLib.HarassHeroBehavior[&quot;Execute&quot;]
-behaviorLib.HarassHeroBehavior[&quot;Execute&quot;] = HarassHeroExecuteOverride
-
-----------------------------------------
---          Retreat Behaviour         --
-----------------------------------------
-local function retreatFromThreatExecuteOverride(botBrain)
-	--note that these actions do not requite a frame to complete, therefore bActionTaken is not necessary.
-	--use energizer when retreating
-	if itemEnergizer and itemEnergizer:CanActivate() and behaviorLib.lastRetreatUtil&gt;=object.nEnergizerThreshold then
-		botBrain:OrderItem(core.itemEnergizer.object or core.itemEnergizer, false)
-		object.nTimeEnergizered=HoN:GetGameTime()
-	end
-	--use regurgitate when retreating
-	if (skills.abilRegurgitate:CanActivate() and behaviorLib.lastRetreatUtil&gt;=object.nRegurgitateThreshold) then
-		core.OrderAbility(botBrain, skills.abilRegurgitate,false,true)
-	end
-	--use ghost marchers when retreating
-	if (core.itemGhostMarchers and core.itemGhostMarchers:CanActivate())then --when heading to base, use boots
-		botBrain:OrderItem(core.itemGhostMarchers.object or core.itemGhostMarchers, false)
-	end
-	
-	return object.RetreatFromThreatExecuteOld(botBrain)
-end
-object.RetreatFromThreatExecuteOld = behaviorLib.RetreatFromThreatBehavior[&quot;Execute&quot;]
-behaviorLib.RetreatFromThreatBehavior[&quot;Execute&quot;] = retreatFromThreatExecuteOverride
-
-----------------------------------------
---      Advanced Think Behaviour      --
-----------------------------------------
-local function advancedThinkUtility(botBrain)
-	return 95; --always ridiculously important. Though, this rarly returns true.
-end
-
-local function advancedThinkExecute(botBrain)
-	local unitSelf = core.unitSelf
-	local vecSelfPos = unitSelf:GetPosition()
-	local nNumMinions=table.getn(object.tMinions)
-	local vecWellPos = core.allyWell and core.allyWell:GetPosition() or behaviorLib.PositionSelfBackUp()
-	local distToWellSq=Vector3.Distance2DSq(vecSelfPos, vecWellPos)
-	local bActionTaken=false
-	
-	--make and eat corpses at base
-	if ( distToWellSq&lt;500*500 and not unitSelf:IsChanneling() and
-	unitSelf:GetLevel()&gt;1 and (skills.abilSpawnMinions:GetCharges()&lt;skills.abilSpawnMinions:GetLevel()*3 or object.nTimeBarfed+4800&gt;HoN:GetGameTime())) then
-		if (skills.abilRegurgitate:CanActivate() and not unitSelf:IsChanneling() and distToWellSq&lt;100*100) then --turn and spew
-			bActionTaken=core.OrderMoveToPos(botBrain,unitSelf, positionOffset(vecSelfPos, 0, 10))
-			object.nTimeBarfed=HoN:GetGameTime()
-			bActionTaken=core.OrderAbility(botBrain, skills.abilRegurgitate,false,true)
-		elseif (object.nTimeBarfed+3300&gt;HoN:GetGameTime()) then --wait while barfing.
-			bActionTaken=core.OrderHoldClamp(botBrain, unitSelf, false)
-		elseif (object.nTimeBarfed+4200&gt;HoN:GetGameTime()) then --collect corpses
-			bActionTaken=core.OrderMoveToPos(botBrain,unitSelf, positionOffset(vecSelfPos, 0, 300))
-		else --move back into the center of the well
-			bActionTaken=core.OrderMoveToPos(botBrain,unitSelf, vecWellPos)
-		end
-	end
-	--remove invalid hatchet item.
-	if core.itemHatchet ~= nil and not core.itemHatchet:IsValid() then
-		core.itemHatchet = nil
-	end
-	--hatchet
-	if not bActionTaken and core.unitCreepTarget and core.itemHatchet and core.itemHatchet:CanActivate() and --can activate
-	  Vector3.Distance2DSq(unitSelf:GetPosition(), core.unitCreepTarget:GetPosition()) &lt;= 600*600 and --in range of hatchet.
-	  unitSelf:GetBaseDamage()*(1-core.unitCreepTarget:GetPhysicalResistance())&gt;core.unitCreepTarget:GetHealth() and --low enough hp, killable
-	  string.find(core.unitCreepTarget:GetTypeName(), &quot;Creep&quot;) then-- viable creep (this makes it ignore minions etc, some of which aren't hatchetable.)
-		botBrain:OrderItemEntity(core.itemHatchet.object or core.itemHatchet, core.unitCreepTarget.object or core.unitCreepTarget, false)--use hatchet.
-	end
-	
-	return bActionTaken
-end
-behaviorLib.advancedThink = {}
-behaviorLib.advancedThink[&quot;Utility&quot;] = advancedThinkUtility
-behaviorLib.advancedThink[&quot;Execute&quot;] = advancedThinkExecute
-behaviorLib.advancedThink[&quot;Name&quot;] = &quot;advancedThink&quot;
-tinsert(behaviorLib.tBehaviors, behaviorLib.advancedThink)
-
-----------------------------------------
---     Collect Corpses Behaviour      --
-----------------------------------------
-local corpsePosition=nil
-local function getCorpsesUtility(botBrain)
-	local unitSelf=core.unitSelf
-	local vecSelfPos=unitSelf:GetPosition()
-	
-	-- Find Corpses
-	if (unitSelf:IsAlive() and core.localUnits~=nil)then
-		local localCreeps={}
-		--Get all creeps nearby and put them into a single table.
-		for k,v in pairs(core.localUnits['EnemyCreeps']) do localCreeps[k] = v end
-		for k,v in pairs(core.localUnits['AllyCreeps']) do localCreeps[k] = v end
-		for k,v in pairs(localCreeps) do object.tOldLocalCreeps[k] = nil end
-		for key, unit in pairs(object.tOldLocalCreeps) do
-			id=key
-			if (unit and unit:IsValid() and unit:GetHealth() and unit:GetHealth()&lt;=0 and string.find(unit:GetTypeName(), &quot;Creep&quot;)) then --these cause awful amounts of 'corpses' constantly.
-				if (unit:GetAttackRange()~=690) then--catapults don't have corpses..
-					table.insert(object.tCorpseTable,{position=unit:GetPosition(),time=HoN:GetGameTime()})--create corpse record
-				end
-			end
-		end
-		object.tOldLocalCreeps=localCreeps
-		for key, event in pairs(object.tCorpseTable) do --remove old/obtained corpses
-			if (HoN.GetGameTime()&gt;event['time']+14000 or Vector3.Distance2DSq(vecSelfPos,event['position'])&lt;=105*105) then
-				table.remove(object.tCorpseTable,key)
-			end
-		end
-	end
-
-	--assess corpses
-	local bDebugLines=false
-	if (unitSelf:GetLevel()&lt;2 or (skills.abilSpawnMinions:GetCharges()==skills.abilSpawnMinions:GetLevel()*3 and unitSelf:GetHealthPercent()*1000==1000))then return 0 end
-	local closestCorpse = nil
-	local nClosestCorpseDistSq = 9999*9999
-	for key, v in pairs(object.tCorpseTable) do
-		vecCorpsePosition = v['position']
-		--&quot;safe&quot; corpses aren't toward the opponents.
-		if not object.vecLaneForward or abs(core.RadToDeg(core.AngleBetween(vecCorpsePosition - unitSelf:GetPosition(), -object.vecLaneForward)) ) &lt; 130 then
-			local nDistSq = Vector3.Distance2DSq(vecCorpsePosition, unitSelf:GetPosition())
-			if (bDebugLines)then core.DrawXPosition(vecCorpsePosition, 'aqua') end
-			if nDistSq &lt; nClosestCorpseDistSq then
-				closestCorpse = v
-				nClosestCorpseDistSq = nDistSq
-			end
-		else if (bDebugLines)then core.DrawXPosition(vecCorpsePosition, 'black') end end
-	end
-	if (closestCorpse) then
-		corpsePosition=closestCorpse['position']
-		return 30*((500-sqrt(nClosestCorpseDistSq))/500) -- basically, the closer you are to the corpse, the more you want it. 30 utility over 500 units.
-	else
-		corpsePosition=nil
-		return 0
-	end
-end
-local function getCorpsesExecute(botBrain)
-	if (corpsePosition) then --corpse exists
-		return core.OrderMoveToPosClamp(botBrain,core.unitSelf, corpsePosition)--get the corpse.
-	end
-	return false
-end
-behaviorLib.getCorpses = {}
-behaviorLib.getCorpses[&quot;Utility&quot;] = getCorpsesUtility
-behaviorLib.getCorpses[&quot;Execute&quot;] = getCorpsesExecute
-behaviorLib.getCorpses[&quot;Name&quot;] = &quot;getCorpses&quot;
-tinsert(behaviorLib.tBehaviors, behaviorLib.getCorpses)
- 
- --@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ PERSONALITY
-core.tKillChatKeys={
-	&quot;*Spews all over keyboard*&quot;,
-	&quot;Get trashed kid.&quot;,
-	&quot;BWAHAHAHAHA!&quot;,
-	&quot;Yummy!&quot;,
-	&quot;Feeling.. trapped?&quot;
-}
-core.tDeathChatKeys = {
-	&quot;Going to leave me here to rot?&quot;,
-	&quot;Bargh, my minions got stuck..&quot;,
-	&quot;You make me sick!&quot;,
-	&quot;I got S2'd..&quot;,
-	&quot;Team? TEAM??&quot;,
-	&quot;Hey, you used a cheat code!&quot;,
-	&quot;Go my minions!&quot;
-} 
- 
- --@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ SIGN OFF
-BotEcho('finished loading balfagore_main')</textarea>
-		</div>			
-	</form>
-	<script type="text/javascript" language="javascript">
-	$(document).ready(function(){
-	$('textarea').autoResize({minHeight: 80,maxHeight: 250});
-	geturl = "http://pastebin.com/w7fj3tgY";
-		$.getJSON("http://graph.facebook.com/fql?q=SELECT total_count FROM link_stat WHERE url='http://pastebin.com/w7fj3tgY'",
-	function(data) {
-	$('#b_facebook').append(data.data[0].total_count);
-	}); 
-	$.getJSON('http://urls.api.twitter.com/1/urls/count.json?url='+geturl+'&callback=?',
-	function(data) {
-	$('#b_twitter').append(data.count);
-	});
-	})
-    </script>						<div style="margin:10px 0;clear:left"></div>
-					</div>
-					<div class="frame_spacer"><!-- --></div>
-					<div id="footer_top" style="clear:both">	
-						<div class="footer_top_title">Pastebin.com Tools &amp; Applications</div>
-						<div class="footer_top_text">
-							<img src="/i/t.gif" alt="" class="icon24 iphone" /><a href="/tools#iphone">iPhone/iPad</a>
-							<img src="/i/t.gif" alt="" class="icon24 windows" /><a href="/tools#windows">Windows</a>
-							<img src="/i/t.gif" alt="" class="icon24 firefox" /><a href="/tools#firefox">Firefox</a>
-							<img src="/i/t.gif" alt="" class="icon24 chrome" /><a href="/tools#chrome">Chrome</a>
-							<img src="/i/t.gif" alt="" class="icon24 webos" /><a href="/tools#webos">WebOS</a>
-							<img src="/i/t.gif" alt="" class="icon24 android" /><a href="/tools#android">Android</a>
-							<img src="/i/t.gif" alt="" class="icon24 macos" /><a href="/tools#macos">Mac</a>
-							<img src="/i/t.gif" alt="" class="icon24 opera" /><a href="/tools#opera">Opera</a>
-							<img src="/i/t.gif" alt="" class="icon24 clickto" /><a href="/tools#clickto">Click.to</a>
-							<img src="/i/t.gif" alt="" class="icon24 unix" /><a href="/tools#pastebincl">UNIX</a>
-							<img src="/i/t.gif" alt="" class="icon24 windowsphone" /><a href="/tools#windowsphone">WinPhone</a>
-						</div>
-					</div>
-				</div>
-			</div>
-		</div>
-		<div id="footer">
-			<div id="logo_small"></div>
-			<div id="footer_links">
-				<a href="/">create new paste</a> &nbsp;|&nbsp; <a href="/api">api</a> &nbsp;|&nbsp; <a href="/trends">trends</a> &nbsp;|&nbsp; <a href="/users">users</a> &nbsp;|&nbsp; <a href="/faq">faq</a> &nbsp;|&nbsp; <a href="/tools">tools</a> &nbsp;|&nbsp; <a href="/privacy">privacy</a> &nbsp;|&nbsp; <a href="/cookies_policy">cookies policy</a> &nbsp;|&nbsp; <a href="/contact">contact</a> &nbsp;|&nbsp; <a href="/stats">stats</a> &nbsp;|&nbsp; <a href="/pro">go pro</a> 
-				<br />Follow us: <a href="http://www.facebook.com/pages/Pastebincom/150549571626327" target="_blank">pastebin on facebook</a> &nbsp;|&nbsp; <a href="http://twitter.com/#!/pastebin" target="blank">pastebin on twitter</a> &nbsp;|&nbsp; <a href="https://www.google.com/search?gl=us&amp;pz=1&amp;cf=all&amp;ned=us&amp;hl=en&amp;tbm=nws&amp;as_oq=pastebin&amp;as_occt=any&amp;as_qdr=d&amp;authuser=0" target="_blank">pastebin in the news</a>
-
-
-				<br />Pastebin v3.1 rendered in: 0.021 seconds				
-			</div>
-			<div id="footer_right" title="DDoS Protection and Dedicated Server provided by Gigenet.com" onclick="location.href='http://www.gigenet.com/'"></div>
-		</div>
-		<script type="text/javascript">
-		$('.narrow_it').click(function(){
-		    $('#super_frame').animate({width:'100%'}, 500);
-		    $('#footer').animate({width:'100%'}, 500);
-			$(".narrow_it").hide();
-			$(".wide_it").show();
-			$.get('/layout.php', function(data) {
-			});
-		});
-		$('.wide_it').click(function(){
-		    $('#super_frame').animate({width:'1200px'}, 500);
-		    $('#footer').animate({width:'1200px'}, 500);
-			$(".wide_it").hide();
-			$(".narrow_it").show();
-			$.get('/layout.php', function(data) {
-			});
-		});
-		</script>
-
-	</body>
-</html>
